### 2020-02-17 #2098 @Luthaf
After looking through the list, I think mnf2014, nmoran, jochym, trallard, and utf have experience in related fields.

Additionally, b-freitas, kblin, leonardojaneis and zoomelectrico have experience with the web platform/javascript; which is the implementation language of chemiscope.
