### 2020-02-24 #2117 @whedon
[REVIEW]: Chemiscope: interactive structure-property explorer for materials and molecules
**Submitting author:** @luthaf (<a href="http://orcid.org/0000-0003-4824-6512">Guillaume Fraux</a>)
**Repository:** <a href="https://github.com/cosmo-epfl/chemiscope" target ="_blank">https://github.com/cosmo-epfl/chemiscope</a>
**Version:** 0.2.0
