### 2020-02-24 #2117 @mkhorton
Hi @Luthaf, I'm trying the interactive version on chemiscope.org and the Jmol viewer doesn't seem to be showing anything, looks like it should be an easy fix:
