### 2020-02-25 e940889
Some Safari desktop fixes
