### 2020-02-25 #2117 @Luthaf
Thanks for the report @mkhorton, it should be fixed now. If not, feel free to open more issues on chemiscope repo!
