### 2020-03-11 61ef488
Improve error message in sanity check for map data
