### 2020-03-12 f44f3d2
changed typing error in the Using the Standalone Visualizer section: correspinding to corresponding (#14)
