### 2020-03-12 e2012c3
Make 'extra' optional in write_chemiscope_input
