### 2020-03-13 4b208b5
Don't try to get env index from viewer when in structure mode
