### 2020-03-30 fd11e07
Setup tslint and make it happy
