+++ src/structure/index.ts
@@ -1,7 +1,7 @@
 /**
  * This module contains the [[StructureViewer|structure viewer]] class and
  * associated code. Notably, it contains the [[JSmolWidget]] class, that can be
- * used outside of chemiscope as a small wrapper around JSmol.  
+ * used outside of chemiscope as a small wrapper around JSmol.
  *
  * @packageDocumentation
  * @module structure
