+++ src/utils/index.ts
@@ -4,12 +4,15 @@
  */
 
 import {makeDraggable} from './draggable';
-import {Indexes, EnvironmentIndexer, DisplayMode} from './indexer';
-import {sendWarning, addWarningHandler} from './warnings';
+import {DisplayMode, EnvironmentIndexer, Indexes} from './indexer';
+import {addWarningHandler, sendWarning} from './warnings';
 
 function generateGUID() {
-    return "xxxxxxxx-xxxx-4xxx-yxxx-xxxxxxxxxxxx".replace(/[xy]/g, (c) => {
-        var r = Math.random() * 16 | 0, v = c === "x" ? r : (r & 0x3 | 0x8);
+    return 'xxxxxxxx-xxxx-4xxx-yxxx-xxxxxxxxxxxx'.replace(/[xy]/g, (c) => {
+        // tslint:disable-next-line:no-bitwise
+        const r = Math.random() * 16 | 0;
+        // tslint:disable-next-line:no-bitwise
+        const v = c === 'x' ? r : (r & 0x3 | 0x8);
         return v.toString(16);
     });
 }
@@ -32,4 +35,4 @@ export {
     Indexes,
     DisplayMode,
     EnvironmentIndexer,
-}
+};

[…]
