### 2020-03-31 f930da0
Allow to dynamically load structures to reduce the inital loading time
