### 2020-03-31 66f566a
Workaround transparency bug(s) in 2D and 3D mode
