+++ src/map/map.ts
@@ -8,7 +8,7 @@ import assert from 'assert';
 import {Property} from '../dataset';
 import {EnvironmentIndexer, getByID, Indexes, makeDraggable} from '../utils';
 
-import * as Plotly from './plotly/plotly-scatter';
+import Plotly from './plotly/plotly-scatter';
 import {Config, Data, Layout, PlotlyScatterElement} from './plotly/plotly-scatter';
 
 import {COLOR_MAPS} from './colorscales';
@@ -33,13 +33,6 @@ const DEFAULT_LAYOUT = {
         colorscale: [] as Plotly.ColorScale,
         showscale: true,
     },
-    // coloraxis2 is used for the markers lines
-    coloraxis2: {
-        cmax: 0,
-        cmin: 0,
-        colorscale: [] as Plotly.ColorScale,
-        showscale: false,
-    },
     hovermode: 'closest',
     legend: {
         itemclick: false,
@@ -57,8 +50,7 @@ const DEFAULT_LAYOUT = {
     scene: {
         camera: {
             projection: {
-                // broken, see https://github.com/plotly/plotly.js/issues/4514
-                // type: 'orthographic'
+                type: 'orthographic',
             },
         },
         xaxis: {
@@ -546,8 +538,6 @@ export class PropertiesMap {
                     'coloraxis.cmin': min,
                     'coloraxis.colorbar.title.text': this._current.color,
                     'coloraxis.showscale': true,
-                    'coloraxis2.cmax': max,
-                    'coloraxis2.cmin': min,
                 } as unknown as Layout);
 
             } else {
@@ -567,7 +557,6 @@ export class PropertiesMap {
             this._restyle({
                 'hovertemplate': this._hovertemplate(),
                 'marker.color': this._colors(0),
-                'marker.line.color': this._lineColors(0),
             } as Data, 0);
         };
 
@@ -577,8 +566,6 @@ export class PropertiesMap {
             this._relayout({
                 'coloraxis.cmax': max,
                 'coloraxis.cmin': min,
-                'coloraxis2.cmax': max,
-                'coloraxis2.cmin': min,
             } as unknown as Layout);
         };
         this._settings.color.min.onchange = colorRangeChange;
@@ -592,8 +579,6 @@ export class PropertiesMap {
             this._relayout({
                 'coloraxis.cmax': max,
                 'coloraxis.cmin': min,
-                'coloraxis2.cmax': max,
-                'coloraxis2.cmin': min,
             } as unknown as Layout);
         };
 
@@ -602,7 +587,6 @@ export class PropertiesMap {
             this._current.colorscale = this._settings.palette.value;
             this._relayout({
                 'coloraxis.colorscale': this._colorScale(),
-                'coloraxis2.colorscale': this._lineColorScale(),
             } as unknown as Layout);
         };
 
@@ -756,7 +740,6 @@ export class PropertiesMap {
                 coloraxis: 'coloraxis',
                 line: {
                     color: lineColors[0],
-                    coloraxis: 'coloraxis2',
                     width: 1,
                 },
                 // prevent plolty from messing with opacity when doing bubble
@@ -786,7 +769,7 @@ export class PropertiesMap {
                 color: colors[1],
                 line: {
                     color: lineColors[1],
-                    width: 0.5,
+                    width: 1,
                 },
                 size: sizes[1],
             },
@@ -830,9 +813,6 @@ export class PropertiesMap {
         layout.coloraxis.colorscale = this._colorScale();
         layout.coloraxis.cmin = parseFloat(this._settings.color.min.value);
         layout.coloraxis.cmax = parseFloat(this._settings.color.max.value);
-        layout.coloraxis2.cmin = parseFloat(this._settings.color.min.value);
[…]
