### 2020-04-01 fa7ff3c
Update download-example-input script
