### 2020-04-01 e95d528
Do not update selected env on double click
