### 2020-04-06 6a56657
Allow giving the loadStructure callback in the intial config
The structure viewer was trying to load structure 0 before the user could set
the callback
