### 2020-04-14 482a86c
Small paper improvements
