+++ paper/paper.md
@@ -44,25 +44,24 @@ creating 2D or 3D maps of the full database.
 
 ![The Qm7b database [@Montavon2013] visualized with chemiscope](screenshot.png)
 
-Chemiscope is an graphical tool for the interactive exploration of materials and
+Chemiscope is a graphical tool for the interactive exploration of materials and
 molecular databases, correlating local and global structural descriptors with
 the physical properties of the different systems. The interface is composed by
 two panels. The left panel consists of a 2D or 3D scatter plot, in which each
-point corresponds to a chemical entity. The axes, color, style of each point can
-be set to represent a property or a structural descriptor, so that the map makes
-it possible to visualize the structure-property relations. Structural
-descriptors are not computed directly by chemiscope, but must be obtained from
-one of the many codes implementing such descriptors [@librascal; @QUIP].
-Since many of the most common descriptors can be very high dimensional, it can
-also be convenient to apply a dimensionality reduction algorithm that maps them
-to a lower-dimensional space. For example the sketchmap algorithm
-[@Ceriotti2011] was used with the Smooth Overlap of Atomic Positions descriptor
-[@Bartok2013] to generate the visualization above. The right panel consists of a
-visualization of the three-dimensional structure of the chemical entities,
-including possibly the periodic repeat unit, which can be non orthorhombic, and
-of a slider that runs through the sequence of entities in the database.
-Visualizing the chemical structure can help finding an intuitive rationalization
-of the layout of the dataset and the structure-property relations.
+point corresponds to a chemical entity. The axes, color, and style of each point
+can be set to represent a property or a structural descriptor to visualize
+structure-property relations directly. Structural descriptors are not computed
+directly by chemiscope, but must be obtained from one of the many codes
+implementing such descriptors [@librascal; @QUIP]. Since the most common
+descriptors can be very high dimensional, it can be convenient to apply a
+dimensionality reduction algorithm that maps them to a lower-dimensional space
+for easier visualization. For example the sketchmap algorithm [@Ceriotti2011]
+was used with the Smooth Overlap of Atomic Positions descriptor [@Bartok2013] to
+generate the visualization above. The right panel displays the three-dimensional
+structure of the chemical entities, possibly including periodic repetition for
+crystals. Visualizing the chemical structure can help finding an intuitive
+rationalization of the layout of the dataset and the structure-property
+relations.
 
 Whereas similar tools [@Gong2013; @Gutlein2014; @Probst2017; @ISV] only allow
 visualizing maps and structures in which each data point corresponds to a
@@ -94,20 +93,19 @@ datasets containing up to a million points, reacting to user input within a few
 hundred milliseconds in the default 2D mode. More elaborate visualization are
 slower, while still handling 100k points easily.
 
-The web platform makes chemiscope usable from different operating systems
-without the need to develop and maintain separate ports. It also makes it easy
-to provide an online service that allows users without installation to visualize
-their own dataset. One instance of this online service is currently available at
-http://chemiscope.org. Chemiscope is implemented as library of re-usable
+The use of web technologies makes chemiscope usable from different operating
+systems without the need to develop, maintain and package the code for each
+operating system. It also mean that we can provide an online service at
+http://chemiscope.org allowing users to visualize their own dataset without
+installing anything. Chemiscope is implemented as library of re-usable
 components linked together via callbacks. This makes it easy to modify the
-default interface to generate mode elaborate visualizations: it would be simple,
-for example, to re-use the molecular viewer to display multiple maps generated
-with different parameters of a dimensionality reduction algorithm. Chemiscope
-can also be distributed in a standalone mode, where the code and a predefined
-dataset are merged together as a single HTML file. This standalone mode is
[…]
