### 2020-04-14 4fb0503
Display properties on top of the structure viewer
