### 2020-04-16 c654144
Ignore double click detection in 3D mode
Plolty is not passing the native event in this case
