### 2020-04-16 2adac1e
Update the selected atom in structure viewer when the user click there
