### 2020-04-17 dc4783f
Hide selected marker if x or y are NaN or INF
