### 2020-04-20 49a4935
Fix arrayMaxMin
Number.MIN_VALUE is not the smallest value, it is the smaller positive value than
can represetented
