### 2020-05-01 6adbbe2
Don't really know what I'm doing, but this triggers the reupdate of the color scheme
Changing only the range doesn't seem to refresh the plotly panel
