+++ src/map/map.ts
@@ -567,8 +567,10 @@ export class PropertiesMap {
             this._relayout({
                 'coloraxis.cmax': max,
                 'coloraxis.cmin': min,
+                 'coloraxis.colorscale': this._colorScale(),
             } as unknown as Layout);
         };
+        
         this._settings.color.min.onchange = colorRangeChange;
         this._settings.color.max.onchange = colorRangeChange;
 
