### 2020-05-01 b5705f0
Just created a stub of a "preset" option for a dataset
