### 2020-05-04 8ee71b9
Also fix color scale reset
