### 2020-05-04 cad6d9e
Trigger the update of the plotly color scheme (#18)
* Don't really know what I'm doing, but this triggers the reupdate of the color scheme

Changing only the range doesn't seem to refresh the plotly panel

* Also fix color scale reset

Co-authored-by: Luthaf <luthaf@luthaf.fr>
