+++ src/map/map.ts
@@ -567,6 +567,7 @@ export class PropertiesMap {
             this._relayout({
                 'coloraxis.cmax': max,
                 'coloraxis.cmin': min,
+                'coloraxis.colorscale': this._colorScale(),
             } as unknown as Layout);
         };
         this._settings.color.min.onchange = colorRangeChange;
@@ -580,6 +581,7 @@ export class PropertiesMap {
             this._relayout({
                 'coloraxis.cmax': max,
                 'coloraxis.cmin': min,
+                'coloraxis.colorscale': this._colorScale(),
             } as unknown as Layout);
         };
 
