### 2020-05-04 e4afee3
Access git version and display it
