### 2020-05-04 fb1abbe
Allow to use symbols in 3D mode
