+++ src/map/map.ts
@@ -6,7 +6,7 @@
 import assert from 'assert';
 
 import {Property} from '../dataset';
-import {EnvironmentIndexer, getByID, Indexes, makeDraggable} from '../utils';
+import {EnvironmentIndexer, getByID, Indexes, makeDraggable, sendWarning} from '../utils';
 
 import Plotly from './plotly/plotly-scatter';
 import {Config, Data, Layout, PlotlyScatterElement} from './plotly/plotly-scatter';
@@ -101,6 +101,15 @@ const DEFAULT_CONFIG = {
         'resetCameraLastSave3d',
     ],
 };
+
+// in 3D mode, only strings are supported for 'marker.symbol', and only very few
+// of them. See https://github.com/plotly/plotly.js/issues/4205 as the plotly
+// issue tracking more symbols in 3D mode.
+const POSSIBLE_SYMBOLS_IN_3D = ['circle', 'square', 'diamond', 'cross', 'x'];
+
+function get3DSymbol(i: number): string {
+    return POSSIBLE_SYMBOLS_IN_3D[i % POSSIBLE_SYMBOLS_IN_3D.length];
+}
 
 // get the max/min of an array. Math.min(...array) fails with very large arrays
 function arrayMaxMin(values: number[]): {max: number, min: number} {
@@ -972,14 +981,25 @@ export class PropertiesMap {
      * Get the values to use as marker symbol with the given plotly `trace`, or
      * all of them if `trace === undefined`.
      */
-    private _symbols(trace?: number): Array<number | number[]> {
+    private _symbols(trace?: number): Array<number | number[] | string | string[]> {
         if (this._current.symbols !== undefined) {
-            assert(!this._is3D());
             const values = this._properties()[this._current.symbols].values;
-            return this._selectTrace<number | number[]>(values, values[this._selected], trace);
+            if (this._is3D()) {
+                // If we need more symbols than available, we'll send a warning
+                // and repeat existing ones
+                if (this._symbolsCount() > POSSIBLE_SYMBOLS_IN_3D.length) {
+                    sendWarning(`${this._symbolsCount()} symbols are required, but we only have ${POSSIBLE_SYMBOLS_IN_3D.length}. Some symbols will be repeated`);
+                }
+                const symbols = values.map(get3DSymbol);
+                return this._selectTrace<string | string[]>(symbols, symbols[this._selected], trace);
+            } else {
+                // in 2D mode, use automatic assignment of symbols from numeric
+                // values
+                return this._selectTrace<number | number[]>(values, values[this._selected], trace);
+            }
         } else {
             // default to 0 (i.e. circles)
-            return this._selectTrace<number | number[]>(0, 0, trace);
+            return this._selectTrace<string | string[]>('circle', 'circle', trace);
         }
     }
 
@@ -988,7 +1008,6 @@ export class PropertiesMap {
         const result = [false, false];
 
         if (this._current.symbols !== undefined) {
-            assert(!this._is3D());
             for (let i = 0; i < this._symbolsCount()!; i++) {
                 result.push(true);
             }
@@ -1006,7 +1025,6 @@ export class PropertiesMap {
         const result = ['', ''];
 
         if (this._current.symbols !== undefined) {
-            assert(!this._is3D());
             const names = this._properties()[this._current.symbols].string!.strings();
             for (const name of names) {
                 result.push(name);
@@ -1065,18 +1083,19 @@ export class PropertiesMap {
     /** Switch current plot from 2D to 3D */
     private _switch3D() {
         assert(this._is3D());
-        // Can not use symbols with 3D plots, why?
-        this._settings.symbol.disabled = true;
-        this._current.symbols = undefined;
         this._settings.z.scale.disabled = false;
         this._settings.z.min.disabled = false;
[…]
