### 2020-05-04 ea51183
Use sizemode: area to ensure consistent sizing of symbols in 3D mode
