### 2020-05-07 #2117 @mkhorton
Hi @Luthaf, first my sincere apologies in the delay in making significant progress on this review. The code looks excellent, very well structured, and fills a clear need (indeed, we work on similar problems ourselves, and it's been a common source of frustration that all these nice global similarity measures exist, but are pretty inaccessible to the average user).

I have two queries:

1. Are there any automated tests? I couldn't find any.

2. Could you improve the paragraph for people looking to contribute and/or where to find support, just a minor point.
