### 2020-05-07 #2117 @Luthaf
> Are there any automated tests? I couldn't find any.

No, unfortunately, I could not find a nice way to do automated tests for web user interfaces. If you know of a good framework for this, I am interested. I'd like to add such tests at some point.

In the mean time, I have been using the example files (which you can get locally with `npm run  download-example-input`) to test the behaviour of the code on varied examples.

> Could you improve the paragraph for people looking to contribute and/or where to find support, just a minor point.

Sure!
