### 2020-05-08 #2117 @mkhorton
>  If you know of a good framework for this, I am interested. I'd like to add such tests at some point.

Yes, there are many test frameworks available. It is a bit of a pain to write tests for web apps, but it's worth it. Even minimal tests that just check your server loads and displays a graph and molecule can be valuable if, for example, something like a dependency being updated breaks your app.

Some frameworks include [Jest](https://jestjs.io) for testing components, browser testing like [Selenium](https://www.selenium.dev) or visual diffing like [Percy](https://percy.io) (commercial).

Automated testing is part of the requirements for JOSS paper acceptance.
