### 2020-05-08 #2117 @Luthaf
> Yes, there are many test frameworks available. It is a bit of a pain to write tests for web apps, but it's worth it. Even minimal tests that just check your server loads and displays a graph and molecule can be valuable if, for example, something like a dependency being updated breaks your app.
>
> Some frameworks include Jest for testing components, browser testing like Selenium or visual diffing like Percy (commercial).

Percy sounds interesting, but I would rather use something free and open source, even if the setup is a bit more complex. However I disagree that minimal tests are useful: when updating dependencies I will be manually checking that basics still work. IMO, more interesting tests would be workflow/user story tests (when I click on this element, this other element changes, and this modal opens); do you know a framework for that?

> Automated testing is part of the requirements for JOSS paper acceptance.

From JOSS [documentation](https://joss.readthedocs.io/en/latest/review_criteria.html#tests)

> Tests
>
>Authors are strongly encouraged to include an automated test suite covering the core functionality of their software.
>
>    Good: An automated test suite hooked up to an external service such as Travis-CI or similar
>    OK: Documented manual steps that can be followed to objectively check the expected functionality of the software (e.g., a sample input file to assert behavior)
>    Bad (not acceptable): No way for you the reviewer to objectively assess whether the software works

I hoped that chemiscope would fell in the OK category here, even if it is not in the Good one.
