### 2020-05-08 d286eae
Improve error message when the size of properties are wrong
