### 2020-05-08 #2117 @mkhorton
> I hoped that chemiscope would fell in the OK category here, even if it is not in the Good one.

Yes, I agree, on the OK criteria I can check it off.

In my role as a reviewer though, I do want to echo the JOSS guidelines and strongly encourage tests. Tests in JavaScript are definitely possible, and there are open source solutions ([Jest](https://jestjs.io) is fully open source). From my own past experience, writing web apps is a lot more brittle than other kinds of apps precisely because they tend to have very deep dependency trees and lots of small packages.

> when updating dependencies I will be manually checking that basics still work.

The intent here is for other contributors to the project too, and help them, help them know when they break things and how to add new features that don't conflict with requirements for the project. Tests help communicate your intent for what you want the code to do.

Your call though -- I agree that this meets the minimal requirements, so I won't hold up the review over it :)
