### 2020-05-08 #2117 @Luthaf
I am already convinced by the usefulness of testing in general, and its availability in JavaScript projects (I used mocha/chai + karma + Travis running on node + 3 browsers in [another project of mine](https://github.com/chemfiles/chemfiles.js) to get >95% line coverage).

My pain point here is that this project does not tend itself to unit tests, and would require setting up more integrated, end-to-end testing. And I would like to have tests that are useful, and don't require changing them every time I tweak the placement of an element. So graphical diff would be interesting to have -- if I can find an open source version of it -- but not the end of the story.

Anyway, I'll spend some more time looking for frameworks for such end-to-end testing!
