### 2020-05-08 #2117 @mkhorton
Ok, that sounds reasonable. And I'm sure that you know the requirements of your project better than I do, so I did not mean to lecture on this point.

> And I would like to have tests that are useful, and don't require changing them every time I tweak the placement of an element.

Hear you on this!
