### 2020-05-20 3f0cebc
Fix write_chemiscope_input with partial metadata
And make metadata optional
