### 2020-05-20 9560a0d
Implement setings linked to HTML DOM elements
