### 2020-05-20 714f9cd
Implement setings linked to HTML DOM elements
