### 2020-05-20 c7d6d83
Use the new settings for JSmolWidget
