+++ src/structure/widget.ts
@@ -7,7 +7,7 @@ import assert from 'assert';
 
 import {JmolObject, JSmolApplet} from 'jsmol';
 
-import {generateGUID, getByID, makeDraggable} from '../utils';
+import {generateGUID, getByID, HTMLSetting, makeDraggable} from '../utils';
 
 import HTML_SETTINGS from './settings.html';
 
@@ -126,50 +126,43 @@ export class JSmolWidget {
     /// Reference to the JSmol applet to be updated with script calls
     private _applet!: JSmolApplet;
     /// Representation options from the HTML side
-    private _options!: {
-        // show bonds if checked
-        bonds: HTMLInputElement,
-        // use space filling representation if checked
-        spaceFilling: HTMLInputElement;
-        // show atoms labels if checked
-        atomLabels: HTMLInputElement;
-        // show unit cell information and lines if checked
-        unitCell: HTMLInputElement;
-        // spin the represenation if checked
-        rotation: HTMLInputElement;
+    private _settings!: {
+        // should we show bonds
+        bonds: HTMLSetting<'boolean'>,
+        // should we use space filling representation
+        spaceFilling: HTMLSetting<'boolean'>;
+        // should we show atoms labels
+        atomLabels: HTMLSetting<'boolean'>;
+        // should we show unit cell information and lines
+        unitCell: HTMLSetting<'boolean'>;
+        /// Is the current unit cell displayed as a packed cell?
+        packedCell: HTMLSetting<'boolean'>;
+        /// number of repetitions in the `a/b/c` direction for the supercell
+        supercell: [HTMLSetting<'int'>, HTMLSetting<'int'>, HTMLSetting<'int'>];
+        // should we spin the represenation
+        rotation: HTMLSetting<'boolean'>;
         // which axis system to use (none, xyz, abc)
-        axes: HTMLSelectElement;
+        axes: HTMLSetting<'string'>;
         // options related to environments
         environments: {
+            // should we display environments & environments options
+            activated: HTMLSetting<'boolean'>;
             // the cutoff value for spherical environments
-            cutoff: HTMLInputElement;
+            cutoff: HTMLSetting<'number'>;
             // which style for atoms not in the environment
-            bgStyle: HTMLSelectElement;
+            bgStyle: HTMLSetting<'string'>;
             // which colors for atoms not in the environment
-            bgColor: HTMLSelectElement;
-            // reset the cutoff to its original value
-            reset: HTMLButtonElement;
-            // enable/disable environments display
-            toggle: HTMLButtonElement;
+            bgColor: HTMLSetting<'string'>;
         }
         // keep the orientation constant when loading a new structure if checked
-        keepOrientation: HTMLInputElement;
+        keepOrientation: HTMLSetting<'boolean'>;
     };
-    /// Supercell related options and data
-    private _supercell!: {
-        /// number of repetitions in the `a` direction for the supercell
-        repeat_a: HTMLInputElement;
-        /// number of repetitions in the `b` direction for the supercell
-        repeat_b: HTMLInputElement;
-        /// number of repetitions in the `c` direction for the supercell
-        repeat_c: HTMLInputElement;
-        /// button used to reset the supercell
-        reset: HTMLButtonElement;
-        /// Is the current unit cell displayed as a packed cell?
-        packed: HTMLInputElement;
-        /// The supercell used to intialize the viewer
-        initial?: [number, number, number];
-    };
+    /// The supercell used to intialize the viewer
+    private _initialSupercell?: [number, number, number];
+    // button to reset the environment cutoff to its original value
+    private _resetEnvCutof!: HTMLButtonElement;
+    // button to reset reset the supercell
+    private _resetSupercell!: HTMLButtonElement;
[…]
