### 2020-05-25 73eecef
Refactor the map to only use HTML elements as settings
