+++ src/map/map.ts
@@ -129,8 +129,8 @@ function arrayMaxMin(values: number[]): {max: number, min: number} {
 
 /** HTML element holding settings for a given axis (x, y, z, color) */
 interface AxisSetting {
-    /// Which values shoudl we use
-    select: HTMLSelectElement;
+    /// Which property should we use for this axis
+    property: HTMLSelectElement;
     /// Which scale (linear/log) should we use
     scale: HTMLSelectElement;
     /// The minimal value for this axis
@@ -162,27 +162,6 @@ export class PropertiesMap {
     private _selected: number;
     /// environment indexer
     private _indexer: EnvironmentIndexer;
-    /// Currently displayed data
-    private _current!: {
-        /// Name of the properties in `this._properties()` used for x values
-        x: string,
-        /// Name of the properties in `this._properties()` used for y values
-        y: string,
-        /// Name of the properties in `this._properties()` used for z values,
-        /// `undefined` for 2D plots
-        z?: string,
-        /// Name of the colorscale to use
-        colorscale: string,
-        /// Name of the properties in `this._properties()` used for color
-        /// values, `undefined` when using the default colors
-        color?: string,
-        /// Name of the properties in `this._properties()` used for size values,
-        /// `undefined` when using the default sizes
-        size?: string,
-        /// Name of the properties in `this._properties()` used for symbols
-        /// values, `undefined` when using the default symbols
-        symbols?: string,
-    };
     /// callback to get the initial positioning of the settings modal. The
     /// callback gets the current placement of the settings as a DOMRect, and
     /// should return top and left positions in pixels, used with
@@ -194,14 +173,15 @@ export class PropertiesMap {
         y: AxisSetting;
         z: AxisSetting;
         color: AxisSetting;
-        colorReset: HTMLButtonElement;
         palette: HTMLSelectElement;
         symbol: HTMLSelectElement;
         size: {
-            select: HTMLSelectElement;
+            property: HTMLSelectElement;
             factor: HTMLInputElement;
         };
     };
+    /// Button used to reset the range of the color axis
+    private _colorReset: HTMLButtonElement;
     /// Marker indicating the position of the latest selected point in 2D mode
     /// Using such div is much faster than trying to restyle the full plot,
     /// especially with more than 100k points. In 3D mode, a separate trace is
@@ -237,43 +217,42 @@ export class PropertiesMap {
         this._selectedMarker.classList.add('chsp-selected');
 
         this._data = new MapData(properties);
-        this._setupDefaults();
 
         this._createSettings();
         this._settings = {
             color: {
                 max: getByID<HTMLInputElement>('chsp-color-max'),
                 min: getByID<HTMLInputElement>('chsp-color-min'),
+                property: getByID<HTMLSelectElement>('chsp-color'),
                 scale: document.createElement('select'),
-                select: getByID<HTMLSelectElement>('chsp-color'),
             },
             x: {
                 max: getByID<HTMLInputElement>('chsp-x-max'),
                 min: getByID<HTMLInputElement>('chsp-x-min'),
+                property: getByID<HTMLSelectElement>('chsp-x'),
                 scale: getByID<HTMLSelectElement>('chsp-x-scale'),
-                select: getByID<HTMLSelectElement>('chsp-x'),
             },
             y: {
                 max: getByID<HTMLInputElement>('chsp-y-max'),
                 min: getByID<HTMLInputElement>('chsp-y-min'),
+                property: getByID<HTMLSelectElement>('chsp-y'),
                 scale: getByID<HTMLSelectElement>('chsp-y-scale'),
-                select: getByID<HTMLSelectElement>('chsp-y'),
             },
             z: {
                 max: getByID<HTMLInputElement>('chsp-z-max'),
                 min: getByID<HTMLInputElement>('chsp-z-min'),
[…]
