### 2020-05-27 815eabe
Added diverging color scales (#22)
