+++ src/map/colorscales.ts
@@ -550,6 +550,141 @@ const TWILIGHT_SHIFTED: RGBColorMap = [
     [1.0000, [ 47,  20,  54]],
 ];
 
+const SEISMIC: RGBColorMap = [
+    [0.0000, [   0,   0,  77]],
+    [0.0159, [   0,   0,  88]],
+    [0.0317, [   0,   0,  99]],
+    [0.0476, [   0,   0, 111]],
+    [0.0635, [   0,   0, 122]],
+    [0.0794, [   0,   0, 133]],
+    [0.0952, [   0,   0, 144]],
+    [0.1111, [   0,   0, 156]],
+    [0.1270, [   0,   0, 167]],
+    [0.1429, [   0,   0, 178]],
+    [0.1587, [   0,   0, 189]],
+    [0.1746, [   0,   0, 200]],
+    [0.1905, [   0,   0, 212]],
+    [0.2063, [   0,   0, 223]],
+    [0.2222, [   0,   0, 234]],
+    [0.2381, [   0,   0, 245]],
+    [0.2540, [   5,   5, 256]],
+    [0.2698, [  21,  21, 256]],
+    [0.2857, [  37,  37, 256]],
+    [0.3016, [  53,  53, 256]],
+    [0.3175, [  69,  69, 256]],
+    [0.3333, [  85,  85, 256]],
+    [0.3492, [ 101, 101, 256]],
+    [0.3651, [ 117, 117, 256]],
+    [0.3810, [ 134, 134, 256]],
+    [0.3968, [ 150, 150, 256]],
+    [0.4127, [ 166, 166, 256]],
+    [0.4286, [ 182, 182, 256]],
+    [0.4444, [ 198, 198, 256]],
+    [0.4603, [ 214, 214, 256]],
+    [0.4762, [ 230, 230, 256]],
+    [0.4921, [ 246, 246, 256]],
+    [0.5079, [ 256, 246, 246]],
+    [0.5238, [ 256, 230, 230]],
+    [0.5397, [ 256, 214, 214]],
+    [0.5556, [ 256, 198, 198]],
+    [0.5714, [ 256, 182, 182]],
+    [0.5873, [ 256, 166, 166]],
+    [0.6032, [ 256, 150, 150]],
+    [0.6190, [ 256, 134, 134]],
+    [0.6349, [ 256, 117, 117]],
+    [0.6508, [ 256, 101, 101]],
+    [0.6667, [ 256,  85,  85]],
+    [0.6825, [ 256,  69,  69]],
+    [0.6984, [ 256,  53,  53]],
+    [0.7143, [ 256,  37,  37]],
+    [0.7302, [ 256,  21,  21]],
+    [0.7460, [ 256,   5,   5]],
+    [0.7619, [ 248,   0,   0]],
+    [0.7778, [ 240,   0,   0]],
+    [0.7937, [ 232,   0,   0]],
+    [0.8095, [ 224,   0,   0]],
+    [0.8254, [ 216,   0,   0]],
+    [0.8413, [ 208,   0,   0]],
+    [0.8571, [ 200,   0,   0]],
+    [0.8730, [ 192,   0,   0]],
+    [0.8889, [ 184,   0,   0]],
+    [0.9048, [ 176,   0,   0]],
+    [0.9206, [ 168,   0,   0]],
+    [0.9365, [ 160,   0,   0]],
+    [0.9524, [ 152,   0,   0]],
+    [0.9683, [ 144,   0,   0]],
+    [0.9841, [ 136,   0,   0]],
+    [1.0000, [ 128,   0,   0]],
+];
+
+const BRG: RGBColorMap = [
+    [0.0000, [   0,   0, 256]],
[…]
