### 2020-05-29 5e6d14c
Use HTMLSettings in the map
