+++ src/utils/settings.ts
@@ -3,6 +3,10 @@
  * @module utils
  */
 
+import assert from 'assert';
+
+import {getByID} from './index';
+
 /**
  * Possible HTML attributes to attach to a setting
  */
@@ -49,6 +53,41 @@ function parse<T extends SettingsType>(type: T, value: string): SettingsValue<T>
     throw Error(`unknown type '${type}' passed to parse`);
 }
 
+/** Possible origins of a setting change: from JS code, or from DOM interaction */
+export type SettingModificationOrigin = 'JS' | 'DOM';
+
+/** A single DOM element that will be synchronized with the javascript value of a setting */
+interface HTMLSettingElement {
+    /** The DOM element that we should check for changes */
+    element: HTMLElement;
+    /** Which atrribute of the element should we check */
+    attribute: Attribute;
+    /** Store the function used to listen on the element to be able to remove it */
+    listener: (e: Event) => void;
+}
+
+/** Multiple settings stored together as (potentially nested) javascript object */
+export interface SettingGroup {
+    [key: string]: HTMLSetting<any> | SettingGroup;
+}
+
+/**
+ * Call the given callback on each setting inside the given SettingGroup
+ *
+ * @param settings group of settings
+ * @param callback callback operating on a single setting
+ */
+export function foreachSetting(settings: SettingGroup, callback: (s: HTMLSetting<any>) => void): void {
+    for (const key in settings) {
+        const element = settings[key];
+        if (element instanceof HTMLSetting) {
+            callback(element);
+        } else {
+            foreachSetting(element as SettingGroup, callback);
+        }
+    }
+}
+
 /**
  * Simple two-way data binding implementation to store settings in chemiscope.
  *
@@ -57,19 +96,20 @@ function parse<T extends SettingsType>(type: T, value: string): SettingsValue<T>
  * linked HTML element, it is automatically updated everywhere.
  */
 export class HTMLSetting<T extends SettingsType> {
-    /** The value of the setting, accessible to JS/TS code */
-    declare public value: SettingsValue<T>;
+    /** the type of the value stored by this setting */
+    public readonly type: T;
+    /** Callback to validate the new value before propagating changes. */
+    public validate: (value: SettingsValue<T>) => void;
     /** Additional callback to run whenever the setting value changes */
-    public onchange: (value: SettingsValue<T>) => void;
+    public onchange: (value: SettingsValue<T>, origin: SettingModificationOrigin) => void;
 
-    // Store the expected type of the value
-    private _type: T;
-    // Store the value itself
+    // storage of the actual value, this is separated from this.value to allow
+    // intercepting set
     private _value: SettingsValue<T>;
-    // Store linked HTML elements: any update to the setting `value` or any of
+    // list of bound HTML elements: any update to the setting `value` or any of
     // the element will change both the settting value and all of the linked
     // elements
-    private _elements: Array<[HTMLElement, Attribute]>;
+    private _boundList: HTMLSettingElement[];
 
     /**
      * Create a new [[HTMLSetting]] containing a value of the given type.
@@ -80,30 +120,23 @@ export class HTMLSetting<T extends SettingsType> {
      * @param value initial value of the setting
      */
     constructor(type: T, value: SettingsValue<T>) {
-        this._type = type;
+        this.type = type;
         this._value = value;
-        this._elements = [];
+        this._boundList = [];
+        this.validate = () => {};
         this.onchange = () => {};
 
-        // uses a Proxy on a frozen object to catch get/set events
         Object.preventExtensions(this);
-        return new Proxy(this, {
-            get: (_, key) => {
-                if (key === 'value') {
-                    return this._value;
-                } else {
-                    return Reflect.get(this, key);
-                }
-            },
-            set: (_, key, v) => {
-                if (key === 'value') {
-                    this._update(v.toString());
-                    return true;
[…]
