### 2020-05-29 5849e83
Use HTMLSettings in the map
