+++ src/map/map.ts
@@ -6,7 +6,8 @@
 import assert from 'assert';
 
 import {Property} from '../dataset';
-import {EnvironmentIndexer, getByID, Indexes, makeDraggable, sendWarning} from '../utils';
+import {EnvironmentIndexer, HTMLSetting, Indexes, SettingGroup, SettingModificationOrigin} from '../utils';
+import {foreachSetting, getByID, makeDraggable, sendWarning} from '../utils';
 
 import Plotly from './plotly/plotly-scatter';
 import {Config, Data, Layout, PlotlyScatterElement} from './plotly/plotly-scatter';
@@ -128,15 +129,50 @@ function arrayMaxMin(values: number[]): {max: number, min: number} {
 }
 
 /** HTML element holding settings for a given axis (x, y, z, color) */
-interface AxisSetting {
+class AxisSetting {
     /// Which property should we use for this axis
-    property: HTMLSelectElement;
+    public property: HTMLSetting<'string'>;
     /// Which scale (linear/log) should we use
-    scale: HTMLSelectElement;
+    public scale: HTMLSetting<'string'>;
     /// The minimal value for this axis
-    min: HTMLInputElement;
+    public min: HTMLSetting<'number'>;
     /// The maximal value for this axis
-    max: HTMLInputElement;
+    public max: HTMLSetting<'number'>;
+
+    constructor(properties: string[], initial: string) {
+        this.max = new HTMLSetting('number', 0);
+        this.min = new HTMLSetting('number', 0);
+
+        this.property = new HTMLSetting('string', initial);
+        this.property.validate = (value) => {
+            if (properties.includes(value) || (initial === '' && value === '')) {
+                return;
+            }
+            throw Error(`invalid property '${value}' for axis`);
+        };
+
+        this.scale = new HTMLSetting('string', 'linear');
+        this.scale.validate = (value) => {
+            if (value === 'linear' || value === 'log') {
+                return;
+            }
+            throw Error(`invalid value '${value}' for axis scale`);
+        };
+    }
+
+    /** Disable auxiliary settings (min/max/scale) related to this axis */
+    public disable() {
+        this.max.disable();
+        this.min.disable();
+        this.scale.disable();
+    }
+
+    /** Enable auxiliary settings (min/max/scale) related to this axis */
+    public enable() {
+        this.max.enable();
+        this.min.enable();
+        this.scale.enable();
+    }
 }
 
 /**
@@ -168,20 +204,21 @@ export class PropertiesMap {
     /// `position: fixed`
     private _settingsPlacement!: (rect: DOMRect) => {top: number, left: number};
     /// Store the HTML elements used for settings
-    private _settings: {
+    private _settings!: {
         x: AxisSetting;
         y: AxisSetting;
         z: AxisSetting;
         color: AxisSetting;
-        palette: HTMLSelectElement;
-        symbol: HTMLSelectElement;
+        palette: HTMLSetting<'string'>;
+        symbol: HTMLSetting<'string'>;
         size: {
-            property: HTMLSelectElement;
-            factor: HTMLInputElement;
+            property: HTMLSetting<'string'>;
+            factor: HTMLSetting<'number'>;
         };
     };
-    /// Button used to reset the range of the color axis
+    /// Button used to reset the range of color axis
     private _colorReset: HTMLButtonElement;
+
     /// Marker indicating the position of the latest selected point in 2D mode
     /// Using such div is much faster than trying to restyle the full plot,
     /// especially with more than 100k points. In 3D mode, a separate trace is
@@ -218,44 +255,11 @@ export class PropertiesMap {
 
         this._data = new MapData(properties);
 
-        this._createSettings();
-        this._settings = {
-            color: {
[…]
