### 2020-05-30 62c4b5f
Command-line version of chemiscope_input
