### 2020-05-30 cfb6a34
Command-line version of chemiscope_input
