+++ utils/chemiscope_input.py
@@ -210,3 +210,54 @@ def write_chemiscope_input(filename, frames, meta=None, extra=None, cutoff=None)
     else:
         with open(filename, 'w') as file:
             json.dump(data, file)
+
+helpstring = """
+    Command-line utility to generate an input for chemiscope  - an interactive 
+    structure-property explorer. Parses an input file containing atomic structures
+    using the ASE I/O module, and converts it into a JSON file that can be loaded in
+    chemiscope. Frame and environment properties must be written in the same file 
+    containing atomic structures: we recommend the extended xyz format, which is 
+    flexible and simple. In all cases, this utility will simply transfer to the
+    JSON input anything that is readable by ASE. 
+"""
+            
+if __name__ == '__main__':
+
+    # command-line execution. requires ASE IO module
+    import argparse
+    try:
+        import ase.io as ase_io
+    except:
+        raise ImportError("chemiscope_input needs ASE modules to parse structure inputs")
+    
+    parser = argparse.ArgumentParser(description=helpstring)
+    parser.add_argument('-i', '--input', type=str, required=True,
+                        help='Input file name.')
+    parser.add_argument('-o', '--output', type=str, 
+                        help='Chemiscope JSON output file.')
+    parser.add_argument('-c', '--cutoff', type=float, 
+                        help='Generate  atom-centred environments with the \
+                              given cutoff.')                        
+    parser.add_argument('--name', default="", type=str,
+                        help='Dataset name.')
+    parser.add_argument('--description', default="", type=str,
+                        help='Dataset description.')
+    parser.add_argument('--authors', nargs='*', type=str, default=[],
+                        help='List of dataset authors.')
+    parser.add_argument('--references', nargs='*', type=str, default=[],
+                        help='List of references for the dataset.')                 
+    args = parser.parse_args()
+
+    # reads filen with ASE    
+    frames = ase_io.read(args.input, ':')
+    
+    # determines output file name automatically if missing
+    output = args.output
+    if args.output is None:
+        output = args.input + "_chemiscope.json"
+    
+    write_chemiscope_input(output, frames, 
+        meta=dict(name=args.name, description=args.description,
+                authors=args.authors, references=args.references),
+        extra=None, cutoff=args.cutoff)
+    
