### 2020-06-03 9a10091
Allow to run lints for command line
