### 2020-06-03 6136374
Run tests & lints on CI
