### 2020-06-03 c79adf3
Add contribution documentation
