### 2020-06-03 90b22fa
Add contribution documentation
