### 2020-06-03 #2117 @Luthaf
Hello @kblin 👋 , do you think you will have time to review this submission in the near future? We are submitting other papers using chemiscope, and we would like to be able to cite the JOSS paper in these.

@cMadan, is there a way for us to get a "pre-reserved" DOI that will resolve to the chemiscope JOSS paper once it is accepted?
