### 2020-06-03 96c591e
Convert download-example-input to typescript
