### 2020-06-03 066f2c3
Fix Python formatting
