### 2020-06-03 58907cc
Move the CLI code to a function
