### 2020-06-03 872227b
Implement review suggestions
