### 2020-06-03 3732b49
Improve autogenerated help message
