+++ utils/chemiscope_input.py
@@ -232,7 +232,9 @@ def main():
     except ImportError:
         raise ImportError("chemiscope_input needs ASE modules to parse structure inputs")
 
-    parser = argparse.ArgumentParser(description=main.__doc__)
+    # Tweak the autogenerated help output to look nicer
+    formatter = lambda prog: argparse.HelpFormatter(prog, max_help_position=22)
+    parser = argparse.ArgumentParser(description=main.__doc__, formatter_class=formatter)
     parser.add_argument('input', type=str,
                         help='input file containing the structures and properties')
     parser.add_argument('-o', '--output', type=str,
