### 2020-06-03 1b04c36
Do not use frame.arrays["numbers"] as a property
