### 2020-06-04 #2117 @kblin
Hi, my apologies.
The paper is on my todo list for this week.

Cheers,
Kai

--
Kai Blin                                           kblin@biosustain.dtu.dk
Researcher / Scientific Software Engineer
DTU Biosustain                                     http://www.biosustain.dtu.dk/
Building 220, room 327
DK - 2800 Kgs. Lyngby
Denmark
mobile: +45 93511306                               twitter: @kaiblin
