### 2020-06-05 29943a3
Use Record<string, unknown> instead of any when validating JavaScript objects
The compiler will help catching accesses to properties without checking that the
property exists first
