+++ src/structure/structure.ts
@@ -5,7 +5,7 @@
 
 import assert from 'assert';
 
-import {Environment, isPositiveInteger, Structure, UserStructure} from '../dataset';
+import {Environment, isPositiveInteger, JsObject, Structure, UserStructure} from '../dataset';
 import {EnvironmentIndexer, Indexes, PositioningCallback} from '../utils';
 import {generateGUID, getByID, getNextColor, sendWarning} from '../utils';
 
@@ -76,22 +76,27 @@ function groupByStructure(n_structures: number, environments?: Environment[]): E
  * issue with `s` if any, or the empty string if `s` looks like a valid
  * structure.
  */
-function checkStructure(s: any): string {
+function checkStructure(s: JsObject): string {
     if (!('size' in s && typeof s.size === 'number' && isPositiveInteger(s.size))) {
         return 'missing "size" in structure';
     }
 
     for (const key of ['names', 'x', 'y', 'z']) {
-        if (!(key in s && s[key].length !== undefined)) {
-            return `missing '${name}' in structure`;
+        if (!(key in s)) {
+            return `missing "${name}" in structure`;
         }
-        if (s[key].length !== s.size) {
-            return `wrong size for '${name}' in structure, expected ${s.size}, got ${s[name].length}`;
+        const array = s[key];
+        if (!Array.isArray(array)) {
+            return `"${name}" must be an array in structure`;
+        }
+
+        if (array.length !== s.size) {
+            return `wrong size for "${name}" in structure, expected ${s.size}, got ${array.length}`;
         }
     }
 
     if ('cell' in s) {
-        if (s.cell.length !== 9) {
+        if (!(Array.isArray(s.cell) && s.cell.length === 9)) {
             return '"cell" must be an array of size 9 in structure';
         }
     }
@@ -330,7 +335,7 @@ export class StructureViewer {
     private _structureForJSmol(index: number): string {
         if (this._cachedStructures[index] === undefined) {
             const s = this.loadStructure(index, this._structures[index]);
-            const check = checkStructure(s);
+            const check = checkStructure(s as unknown as JsObject);
             if (check !== '') {
                 throw Error(
                     `got invalid object as structure: ${check}` + '\n' +
