### 2020-06-05 adfaffc
Use Record<string, unknown> instead of any when validating JavaScript objects
The compiler will help catching accesses to properties without checking that the
property exists first
