### 2020-06-05 c6cfe68
Added command-line options to generate_standalone.py to simplify the workflow to make a dataset into a portable HTML
