+++ utils/generate_standalone.py
@@ -8,6 +8,7 @@ This creates a file that can be distributed in supplementary information
 '''
 import os
 import base64
+import argparse
 
 from bs4 import BeautifulSoup
 import requests
@@ -72,14 +73,44 @@ def inline_ressources(html):
 
 
 def main():
+    """
+    Command-line utility to generate an stand-alone HTML file containing the
+    chemiscope app. The standalone HTML can be loaded in a browser, and used to 
+    read chemiscope JSON inputs. In addition, a chemiscope JSON can be appended
+    to the HTML to have a fully self-contained viewer for a chosen data set. 
+    """
+
+    # Tweak the autogenerated help output to look nicer
+    formatter = lambda prog: argparse.HelpFormatter(prog, max_help_position=22)
+    parser = argparse.ArgumentParser(description=main.__doc__, formatter_class=formatter)
+    parser.add_argument('input', type=str, default="", nargs="?",
+                        help='chemiscope JSON input that is appended to provide default data')
+    parser.add_argument('-o', '--output', type=str, default="",
+                        help='output file name. Defaults to chemiscope.html or the name of the JSON')
+    args = parser.parse_args()
+    
     html = get_html()
     inline_ressources(html)
 
-    with open('standalone.html', 'w') as fd:
+    output_name = args.output
+    if output_name == "":
+        if args.input == "":
+            output_name = "chemiscope.html"
+        else:
+            output_name = os.path.splitext(args.input)[0] + ".html"
+
+    json_data = ""
+    if args.input != "":
+        with open(args.input) as fd:
+            json_data = fd.read()
+
+    with open(output_name, 'w') as fd:
         fd.write(str(html))
         fd.write('<script id=standalone-json-data type="application/json">')
+        fd.write(json_data)
+        
 
-    print('created standalone.html!')
+    print('created '+ output_name +". happy chemiscoping!")
 
 
 if __name__ == '__main__':
