### 2020-06-05 ca899be
Error message added to deal with someone trying to append a gzipped JSON
