### 2020-06-05 042a58b
Create and deploy a standalone viewer
