### 2020-06-05 c5fc25e
Use a full clone on Travis to be able to get the git version
Shallow clones fails as soon as you are too far away from the last tag
