### 2020-06-05 5ad62a7
fixed bug in generating chemiscope_standalone.html
