### 2020-06-05 b508d37
First minimal workflow using GH Actions
