### 2020-06-05 a995657
First minimal workflow using GH Actions
