+++ .github/workflows/tests.yml
@@ -2,7 +2,7 @@ name: Tests & Lints
 on: [push, pull_request]
 
 jobs:
-  build:
+  test:
     runs-on: ubuntu-latest
     steps:
     - uses: actions/checkout@v2
