+++ .github/workflows/tests.yml
@@ -1,5 +1,9 @@
 name: Tests & Lints
-on: [push, pull_request]
+on:
+  push:
+    branches: [master]
+  pull_request:
+    # Check all PR
 
 jobs:
   test:
