### 2020-06-05 0d2faa5
Option to automatically center the selected environment in the viewer
Also rearranged the environment options to make space for the centering switch
