### 2020-06-05 7caf99f
BS accordion to avoid environments hiding properties
