### 2020-06-05 c29b12a
BS accordion to avoid environments hiding properties
