### 2020-06-05 93a75ee
Cleaned up docs
