### 2020-06-06 d6ae006
Initial implementation of presets for the map settings
