+++ src/index.ts
@@ -10,8 +10,8 @@
  *
  * Other organization of the visualization are possible by using the classes
  * responsible for each sub-panel separately, instead of using the
- * [[DefaultVisualizer]]. In this case, users should make sure to finish the
- * plumbing by setting the right callbacks on each element used.
+ * [[DefaultVisualizer]]. In this case, developers should make sure to finish
+ * the plumbing by setting the right callbacks on each element used.
  *
  * @packageDocumentation
  * @module main
@@ -19,7 +19,7 @@
  */
 
 import {EnvironmentInfo} from './info';
-import {PropertiesMap} from './map';
+import {MapPresets, PropertiesMap} from './map';
 import {MetadataPanel} from './metadata';
 import {StructureViewer} from './structure';
 
@@ -35,8 +35,12 @@ require('./static/chemiscope.css');
 export interface Config {
     /** Id of the DOM element to use for the [[MetadataPanel|metadata display]] */
     meta: string;
-    /** Id of the DOM element to use for the [[PropertiesMap|map]] */
-    map: string;
+    map: {
+        /** Id of the DOM element to use for the [[PropertiesMap|map]] */
+        id: string;
+        /** Presets for map settings */
+        presets: MapPresets;
+    };
     /** Id of the DOM element to use for the [[EnvironmentInfo|environment information]] */
     info: string;
     /** Id of the DOM element to use for the [[StructureViewer|structure viewer]] */
@@ -55,8 +59,8 @@ function validateConfig(o: JsObject) {
         throw Error('missing "meta" key in chemiscope config');
     }
 
-    if (!('map' in o && typeof o.map === 'string')) {
-        throw Error('missing "map" key in chemiscope config');
+    if (!('map' in o && 'id' in (o.map as JsObject) && typeof (o.map as JsObject).id === 'string')) {
+        throw Error('missing "map.id" key in chemiscope config');
     }
 
     if (!('info' in o && typeof o.info === 'string')) {
@@ -93,7 +97,7 @@ class DefaultVisualizer {
      * the browser while everything is loading.
      *
      * @param  config  configuration of the visualizer
-     * @param  dataset dataset to load
+     * @param  dataset visualizer input, containing a dataset and optional visualization presets
      * @return         Promise that resolves to a [[DefaultVisualizer]]
      */
     public static load(config: Config, dataset: Dataset): Promise<DefaultVisualizer> {
@@ -109,12 +113,16 @@ class DefaultVisualizer {
     public structure: StructureViewer;
 
     private _indexer: EnvironmentIndexer;
+    // Stores raw input input so we can output it as JSON later
+    private _dataset: Dataset;
 
     // the constructor is private because the main entry point is the static
     // `load` function
     private constructor(config: Config, dataset: Dataset) {
         validateConfig(config as unknown as JsObject);
         validateDataset(dataset as unknown as JsObject);
+
+        this._dataset = dataset;
 
         const mode = (dataset.environments === undefined) ? 'structure' : 'atom';
         this._indexer = new EnvironmentIndexer(mode, dataset.structures, dataset.environments);
@@ -174,6 +182,14 @@ class DefaultVisualizer {
         this.meta.remove();
         this.info.remove();
         this.structure.remove();
+    }
+
+    /**
+     * Dumps the dataset and settings as a JSON string
+     */
+    public dump(): string {
+        const dumpdata = { ...this._dataset, ...{ presets: {map: this.map.dumpPresets()} } };
+        return JSON.stringify(dumpdata);
     }
 }
 
