+++ src/map/settings.ts
@@ -1,0 +1,209 @@
+/**
+ * @packageDocumentation
+ * @module settings
+ */
+
+import assert from 'assert';
+
+import {HTMLSetting, settingsValidator} from '../utils';
+
+export interface AxisPresets {
+    /// Which property should we use for this axis
+    property: string;
+    /// Which scale (linear/log) should we use
+    scale: string;
+    /// The minimal value for this axis
+    min: number;
+    /// The maximal value for this axis
+    max: number;
+}
+
+const AXIS_DEFAULTS: AxisPresets = {
+     max: 0,
+     min : 0,
+     property : '',
+     scale : 'linear',
+};
+
+/** HTML element holding settings for a given axis (x, y, z, color) */
+export class AxisSettings {
+    public property: HTMLSetting<'string'>;
+    public scale: HTMLSetting<'string'>;
+    public min: HTMLSetting<'number'>;
+    public max: HTMLSetting<'number'>;
+
+    constructor(validProperties: string[]) {
+        this.max = new HTMLSetting('number', AXIS_DEFAULTS.max);
+        this.min = new HTMLSetting('number', AXIS_DEFAULTS.min);
+        this.property = new HTMLSetting('string', AXIS_DEFAULTS.property);
+        this.scale = new HTMLSetting('string', AXIS_DEFAULTS.scale);
+
+        this.property.validate = settingsValidator(validProperties, 'axis');
+        this.scale.validate = settingsValidator(['linear', 'log'], 'axis scale');
+    }
+
+    /**
+     * Applies presets, possibly filling in with default values
+     */
+    public applyPresets(presets: Partial<AxisPresets> = {}): void {
+        const initial: AxisPresets = {
+            ...AXIS_DEFAULTS,
+            ...presets,
+        };
+
+        this.max.value = initial.max;
+        this.min.value = initial.min;
+        this.property.value = initial.property;
+        this.scale.value = initial.scale;
+    }
+
+    /**
+     * Dumps presets, in a way that can e.g. be serialized to json
+     */
+    public dumpPresets(): AxisPresets {
+        return {
+            max: this.max.value,
+            min: this.min.value,
+            property: this.property.value,
+            scale: this.scale.value,
+        };
+    }
+
+    /** Disable auxiliary settings (min/max/scale) related to this axis */
+    public disable() {
+        this.max.disable();
+        this.min.disable();
+        this.scale.disable();
+    }
+
+    /** Enable auxiliary settings (min/max/scale) related to this axis */
+    public enable() {
+        this.max.enable();
+        this.min.enable();
+        this.scale.enable();
+    }
+}
+
+export interface MapPresets {
+    x: Partial<AxisPresets>;
+    y: Partial<AxisPresets>;
+    z: Partial<AxisPresets>;
+    color: Partial<AxisPresets>;
+    palette: string;
+    symbol: string;
+    size: {
+        property?: string,
+        factor?: number,
+    };
+}
+
+const MAP_DEFAULTS: MapPresets = {
+    color: AXIS_DEFAULTS,
+    x: AXIS_DEFAULTS,
+    y: AXIS_DEFAULTS,
+    z: AXIS_DEFAULTS,
+
+    palette: 'inferno',
+    size: {
+        factor: 50.0,
+        property: '',
+    },
+    symbol: '',
+};
+
+export class MapSettings {
+    public x: AxisSettings;
+    public y: AxisSettings;
+    public z: AxisSettings;
+    public color: AxisSettings;
+    public palette: HTMLSetting<'string'>;
+    public symbol: HTMLSetting<'string'>;
+    public size: {
+        property: HTMLSetting<'string'>;
+        factor: HTMLSetting<'number'>;
+    };
+
[…]
