### 2020-06-06 de77f22
Initial implementation of presets for the map settings
