+++ src/map/map.ts
@@ -5,17 +5,19 @@
 
 import assert from 'assert';
 
-import {Property} from '../dataset';
-import {EnvironmentIndexer, Indexes} from '../utils';
-import {HTMLSetting, SettingGroup, SettingModificationOrigin} from '../utils';
-import {GUID, PositioningCallback} from '../utils';
-import {enumerate, foreachSetting, getByID, getFirstKey, makeDraggable, sendWarning} from '../utils';
-
 import Plotly from './plotly/plotly-scatter';
 import {Config, Data, Layout, PlotlyScatterElement} from './plotly/plotly-scatter';
 
-import {COLOR_MAPS} from './colorscales';
+import {Property} from '../dataset';
+import {EnvironmentIndexer, Indexes} from '../utils';
+import {SettingGroup, SettingModificationOrigin} from '../utils';
+import {GUID, PositioningCallback} from '../utils';
+import {enumerate, foreachSetting, getByID, getFirstKey, makeDraggable, sendWarning} from '../utils';
+
 import {MapData, NumericProperty} from './data';
+import {AxisSettings, MapPresets, MapSettings} from './settings';
+
+import {COLOR_MAPS} from './colorscales';
 
 import BARS_SVG from '../static/bars.svg';
 import HTML_SETTINGS from './settings.html';
@@ -131,53 +133,6 @@ function arrayMaxMin(values: number[]): {max: number, min: number} {
     return {max, min};
 }
 
-/** HTML element holding settings for a given axis (x, y, z, color) */
-class AxisSetting {
-    /// Which property should we use for this axis
-    public property: HTMLSetting<'string'>;
-    /// Which scale (linear/log) should we use
-    public scale: HTMLSetting<'string'>;
-    /// The minimal value for this axis
-    public min: HTMLSetting<'number'>;
-    /// The maximal value for this axis
-    public max: HTMLSetting<'number'>;
-
-    constructor(properties: string[], initial: string) {
-        this.max = new HTMLSetting('number', 0);
-        this.min = new HTMLSetting('number', 0);
-
-        this.property = new HTMLSetting('string', initial);
-        this.property.validate = (value) => {
-            if (properties.includes(value) || (initial === '' && value === '')) {
-                return;
-            }
-            throw Error(`invalid property '${value}' for axis`);
-        };
-
-        this.scale = new HTMLSetting('string', 'linear');
-        this.scale.validate = (value) => {
-            if (value === 'linear' || value === 'log') {
-                return;
-            }
-            throw Error(`invalid value '${value}' for axis scale`);
-        };
-    }
-
-    /** Disable auxiliary settings (min/max/scale) related to this axis */
-    public disable() {
-        this.max.disable();
-        this.min.disable();
-        this.scale.disable();
-    }
-
-    /** Enable auxiliary settings (min/max/scale) related to this axis */
-    public enable() {
-        this.max.enable();
-        this.min.enable();
-        this.scale.enable();
-    }
-}
-
 /// interface to contain synchronized parameters for marker instances
 interface MarkerData {
     /// color of the marker, synchronized with the StructureViewer where appropriate
@@ -237,18 +192,7 @@ export class PropertiesMap {
     /// environment indexer
     private _indexer: EnvironmentIndexer;
     /// Store the HTML elements used for settings
-    private _settings!: {
-        x: AxisSetting;
-        y: AxisSetting;
-        z: AxisSetting;
-        color: AxisSetting;
-        palette: HTMLSetting<'string'>;
-        symbol: HTMLSetting<'string'>;
-        size: {
-            property: HTMLSetting<'string'>;
-            factor: HTMLSetting<'number'>;
-        };
-    };
+    private _settings!: MapSettings;
[…]
