### 2020-06-06 b76d855
Initial implementation of presets for the map settings
