+++ src/utils/index.ts
@@ -7,7 +7,7 @@ import assert from 'assert';
 
 import {makeDraggable} from './draggable';
 import {DisplayMode, EnvironmentIndexer, Indexes} from './indexer';
-import {foreachSetting, HTMLSetting, SettingGroup, SettingModificationOrigin} from './settings';
+import {foreachSetting, HTMLSetting, SettingGroup, SettingModificationOrigin, settingsValidator} from './settings';
 import {addWarningHandler, sendWarning} from './warnings';
 
 /** Callback type to position an HTML element.
@@ -106,4 +106,5 @@ export {
     HTMLSetting,
     SettingGroup,
     SettingModificationOrigin,
+    settingsValidator,
 };
