### 2020-06-07 87f06db
First commit. Basically just disabled a bunch of checks so that an environment list that is not as long as #atoms can be loaded
