### 2020-06-08 c88b2eb
Simplify reloading a new dataset
