### 2020-06-08 c9c2730
Simplify reloading a new dataset
