### 2020-06-08 1b3e712
Cleaned up settings window.
