### 2020-06-08 e3f37ee
Also remove settings modales on remove()
