### 2020-06-08 c337faa
Also remove settings modales on remove()
