### 2020-06-08 0fe6dca
Allow multiple structure viewer, grouped in a grid
