+++ src/structure/structure.ts
@@ -6,11 +6,55 @@
 import assert from 'assert';
 
 import {Environment, isPositiveInteger, Structure, UserStructure} from '../dataset';
-import {EnvironmentIndexer, getByID, Indexes} from '../utils';
+import {EnvironmentIndexer, generateGUID, getByID, getNextColor, Indexes, sendWarning} from '../utils';
 
 import {structure2JSmol} from './jsmol';
 import {JSmolWidget} from './widget';
 
+import close_svg from './close.svg';
+import duplicate_svg from './duplicate.svg';
+import HTML_SETTINGS from './settings.html';
+
+const MAX_WIDGETS = 9;
+
+/*
+ * Function to return *optimal* arrangement of n widgets.
+ * Defaults to a 4 x ? grid. Hardcoded for simplicity.
+ */
+function bestArrangement(n: number) {
+  switch (n) {
+    case 1:
+    case 2:
+      return {rows: n, columns: 1};
+      break;
+    case 3:
+    case 4:
+      return {rows: 2, columns: 2};
+      break;
+    case 5:
+    case 6:
+      return {rows: 3, columns: 2};
+      break;
+    case 8:
+    case 7:
+    case 9:
+      return {rows: 3, columns: 3};
+      break;
+    case 10:
+    case 11:
+    case 12:
+      return {rows: 4, columns: 3};
+      break;
+    default:
+      return {rows: 4, columns: 4};
+      break;
+
+  }
+}
+
+/*
+ * Needs documentation.
+ */
 function groupByStructure(n_structures: number, environments?: Environment[]): Environment[][] | undefined {
     if (environments === undefined) {
         return undefined;
@@ -40,10 +84,10 @@ function checkStructure(s: any): string {
 
     for (const key of ['names', 'x', 'y', 'z']) {
         if (!(key in s && s[key].length !== undefined)) {
-            return `missing "${name}" in structure`;
+            return `missing '${name}' in structure`;
         }
         if (s[key].length !== s.size) {
-            return `wrong size for "${name}" in structure, expected ${s.size}, got ${s[name].length}`;
+            return `wrong size for '${name}' in structure, expected ${s.size}, got ${s[name].length}`;
         }
     }
 
@@ -56,13 +100,18 @@ function checkStructure(s: any): string {
     return '';
 }
 
+interface WidgetGridData {
+        widget: JSmolWidget;
+        color: string;
+        current: Indexes;
+        }
 /**
  * The [[StructureViewer]] class displays a molecule or a crystal in 3D using
  * [JSmol](http://wiki.jmol.org/index.php/JSmol) for rendering.
  */
 export class StructureViewer {
     /** Callback used when the user select an environment */
-    public onselect: (indexes: Indexes) => void;
+    public onselect: (indexes: Indexes, selectedGUID?: string) => void;
 
     /**
      * Callback used when a new structure should be loaded
@@ -77,8 +126,14 @@ export class StructureViewer {
      */
     public loadStructure: (index: number, structure: any) => Structure;
 
-    /// The JSmolWidget is responsible for displaying the current structure
-    private _widget: JSmolWidget;
+    // GUID of the Active Widget
+    private _active: string;
+    // Map of Widgets GUIDS to their color, widget, and current structure
+    private _widgetMap: Map<string, WidgetGridData>;
+    // List of Widgets GUIDS in the StructureViewer
+    private _GUIDs: string[];
+    // Documentation needed.
+    private _root: HTMLElement;
     /// Playback delay setting
     private _delay: HTMLInputElement;
     /// List of structures in the dataset
@@ -87,9 +142,11 @@ export class StructureViewer {
     private _cachedStructures: string[];
     /// Optional list of environments for each structure
     private _environments?: Environment[][];
[…]
