### 2020-06-08 50c338c
Allow multiple structure viewer, grouped in a grid
