+++ src/map/map.ts
@@ -175,6 +175,11 @@ class AxisSetting {
     }
 }
 
+interface MarkerData {
+    color: string;
+    current: number;
+    marker: HTMLElement;
+}
 /**
  * The [[PropertiesMap]] class displays a 2D or 3D map (scatter plot) of
  * properties in the dataset, using [plotly.js](https://plot.ly/javascript/)
@@ -186,7 +191,7 @@ class AxisSetting {
  */
 export class PropertiesMap {
     /** Callback fired when the plot is clicked and a new point is selected */
-    public onselect: (indexes: Indexes) => void;
+    public onselect: (indexes: Indexes, selectedGUID?: string) => void;
 
     /// HTML root holding the full plot
     private _root: HTMLElement;
@@ -194,8 +199,15 @@ export class PropertiesMap {
     private _plot!: PlotlyScatterElement;
     /// All known properties
     private _data: MapData;
-    /// Index of the currently selected point
-    private _selected: number;
+
+    /// GUID of the currently selected point
+    private _active: string;
+    /// Dictionary of current markers/GUIDs
+    private _markerMap: Map<string, MarkerData>;//{[GUID: string]: MarkerData; };
+
+    /// list of GUIDs
+    private _GUIDs: string[];
+
     /// environment indexer
     private _indexer: EnvironmentIndexer;
     /// callback to get the initial positioning of the settings modal. The
@@ -221,12 +233,6 @@ export class PropertiesMap {
     /// The HTML element containing the settings modal
     private _settingsModal!: HTMLElement;
 
-    /// Marker indicating the position of the latest selected point in 2D mode
-    /// Using such div is much faster than trying to restyle the full plot,
-    /// especially with more than 100k points. In 3D mode, a separate trace is
-    /// used instead.
-    private _selectedMarker: HTMLElement;
-
     /**
      * Create a new [[PropertiesMap]] inside the DOM element with the given HTML
      * `id`
@@ -236,12 +242,21 @@ export class PropertiesMap {
      *                   environments index to structure/atom indexes
      * @param properties properties to be displayed
      */
-    constructor(id: string, indexer: EnvironmentIndexer, properties: {[name: string]: Property}) {
+    constructor(id: string,
+                indexer: EnvironmentIndexer,
+                properties: {[name: string]: Property},
+                starterGuid: string, /// fix me later
+                ) {
         this._indexer = indexer;
-        this._selected = -1;
         this.onselect = () => {};
 
+        /// new parameters
+        this._active = '';
+        this._markerMap = new Map();
+        this._GUIDs = [];
+
         this._root = getByID(id);
+
         if (this._root.style.position === '') {
             this._root.style.position = 'relative';
         }
@@ -251,11 +266,9 @@ export class PropertiesMap {
         this._plot.style.height = '100%';
         this._root.appendChild(this._plot);
 
-        this._selectedMarker = document.createElement('div');
-        this._root.appendChild(this._selectedMarker);
-        this._selectedMarker.classList.add('chsp-selected');
-
         this._data = new MapData(properties);
+        this._addMarker(starterGuid);
+        this.active = starterGuid;
 
         this._insertSettingsHTML();
         this._colorReset = getByID<HTMLButtonElement>('chsp-color-reset');
@@ -267,13 +280,30 @@ export class PropertiesMap {
     }
 
     /** Change the selected environment to the one with the given `indexes` */
-    public select(indexes: Indexes) {
+    public select(indexes: Indexes, selectedGUID: string = this._active) {
         // Plotly.restyle fires the plotly_click event, so ensure we only run
         // the update once.
         // https://github.com/plotly/plotly.js/issues/1025
-        if (indexes.environment !== this._selected) {
[…]
