+++ src/map/map.ts
@@ -202,8 +202,8 @@ export class PropertiesMap {
 
     /// GUID of the currently selected point
     private _active: string;
-    /// Dictionary of current markers/GUIDs
-    private _markerMap: Map<string, MarkerData>;//{[GUID: string]: MarkerData; };
+    /// Map of current markers/GUIDs
+    private _markerMap: Map<string, MarkerData>;
 
     /// list of GUIDs
     private _GUIDs: string[];
@@ -285,25 +285,25 @@ export class PropertiesMap {
         // the update once.
         // https://github.com/plotly/plotly.js/issues/1025
         if (selectedGUID !== undefined) {
-          if (indexes.structure < 0 && indexes.environment < 0) {
-            this._removeMarker(selectedGUID, true);
-          } else {
-            /// Checks if marker exists on this map, if not adds it
-            if (! this._markerMap.has(selectedGUID)) {
-              this._addMarker(selectedGUID, indexes.environment);
-            }
+            if (indexes.structure < 0 && indexes.environment < 0) {
+                this._removeMarker(selectedGUID, true);
+            } else {
+                /// Checks if marker exists on this map, if not adds it
+                if (!this._markerMap.has(selectedGUID)) {
+                    this._addMarker(selectedGUID, indexes.environment);
+                }
 
-            const markerData = this._markerMap.get(selectedGUID);
-            if (markerData !== undefined) {
-              /// Sets the active marker on this map
-              this.active = selectedGUID;
+                const markerData = this._markerMap.get(selectedGUID);
+                if (markerData !== undefined) {
+                    /// Sets the active marker on this map
+                    this.active = selectedGUID;
 
-              if (markerData.current === undefined || indexes.environment !== markerData.current ) {
-                markerData.current = indexes.environment;
-              }
-              this._updateSelectedMarker(selectedGUID);
+                    if (markerData.current === undefined || indexes.environment !== markerData.current ) {
+                        markerData.current = indexes.environment;
+                    }
+                    this._updateSelectedMarker(selectedGUID);
+                }
             }
-          }
         }
     }
 
@@ -1028,7 +1028,7 @@ export class PropertiesMap {
         if (this._settings.symbol.value !== '') {
             const values = this._property(this._settings.symbol.value).values;
             const markerData = this._markerMap.get(this._active);
-            if (markerData !== undefined ) {
+            if (markerData !== undefined) {
               if (this._is3D()) {
                   // If we need more symbols than available, we'll send a warning
                   // and repeat existing ones
@@ -1145,7 +1145,7 @@ export class PropertiesMap {
         } as unknown as Data);
 
         for (const guid of this._GUIDs) {
-          this._removeMarker(guid, false);
+            this._removeMarker(guid, false);
         }
 
         // Change the data that vary between 2D and 3D mode
@@ -1188,10 +1188,10 @@ export class PropertiesMap {
         } as unknown as Data);
 
         for (const guid of this._GUIDs) {
-          const markerData = this._markerMap.get(guid);
-          if (markerData !== undefined ) {
-            this._addMarker(guid, markerData.current);
-          }
+            const markerData = this._markerMap.get(guid);
+            if (markerData !== undefined ) {
+                this._addMarker(guid, markerData.current);
+            }
         }
 
         // Change the data that vary between 2D and 3D mode
@@ -1242,150 +1242,175 @@ export class PropertiesMap {
             }
         }
     }
-    /*
-     * Function to set the active marker for communicating with the structure viewer
+
+    /**
+     * Function to set the active marker for communicating with the structure
+     * viewer
+     *
+     * @param  activeGUID the GUID of the new active viewer
[…]
