### 2020-06-08 085210d
Small formmatting cleanup
