### 2020-06-08 052e1b7
Do not store the GUIDs of selected markers as a separate list
