+++ src/map/map.ts
@@ -202,11 +202,8 @@ export class PropertiesMap {
 
     /// GUID of the currently selected point
     private _active: string;
-    /// Map of current markers/GUIDs
-    private _markerMap: Map<string, MarkerData>;
-
-    /// list of GUIDs
-    private _GUIDs: string[];
+    /// Map associating currently selected markers GUID to additional data
+    private _selected: Map<string, MarkerData>;
 
     /// environment indexer
     private _indexer: EnvironmentIndexer;
@@ -249,11 +246,8 @@ export class PropertiesMap {
                 ) {
         this._indexer = indexer;
         this.onselect = () => {};
-
-        /// new parameters
         this._active = '';
-        this._markerMap = new Map();
-        this._GUIDs = [];
+        this._selected = new Map();
 
         this._root = getByID(id);
 
@@ -289,11 +283,11 @@ export class PropertiesMap {
                 this._removeMarker(selectedGUID, true);
             } else {
                 /// Checks if marker exists on this map, if not adds it
-                if (!this._markerMap.has(selectedGUID)) {
+                if (!this._selected.has(selectedGUID)) {
                     this._addMarker(selectedGUID, indexes.environment);
                 }
 
-                const markerData = this._markerMap.get(selectedGUID);
+                const markerData = this._selected.get(selectedGUID);
                 if (markerData !== undefined) {
                     /// Sets the active marker on this map
                     this.active = selectedGUID;
@@ -854,8 +848,8 @@ export class PropertiesMap {
         });
         this._plot.on('plotly_afterplot', () => this._afterplot());
 
-        for (const guid of this._GUIDs) {
-          this._updateSelectedMarker(guid);
+        for (const guid of this._selected.keys()) {
+            this._updateSelectedMarker(guid);
         }
     }
 
@@ -889,11 +883,8 @@ export class PropertiesMap {
     private _xValues(trace?: number): number[][] {
         const values = this._property(this._settings.x.property.value).values;
         const selected = [];
-        for (const guid of this._GUIDs) {
-          const markerData = this._markerMap.get(guid);
-          if (markerData !== undefined ) {
-            selected.push(values[markerData.current]);
-          }
+        for (const marker of this._selected.values()) {
+            selected.push(values[marker.current]);
         }
 
         if (!this._is3D()) {
@@ -909,12 +900,10 @@ export class PropertiesMap {
     private _yValues(trace?: number): number[][] {
         const values = this._property(this._settings.y.property.value).values;
         const selected = [];
-        for (const guid of this._GUIDs) {
-          const markerData = this._markerMap.get(guid);
-          if (markerData !== undefined ) {
-            selected.push(values[markerData.current]);
-          }
+        for (const marker of this._selected.values()) {
+            selected.push(values[marker.current]);
         }
+
         if (!this._is3D()) {
             selected[0] = NaN;
         }
@@ -932,11 +921,8 @@ export class PropertiesMap {
 
         const values = this._property(this._settings.z.property.value).values;
         const selected = [];
-        for (const guid of this._GUIDs) {
-          const markerData = this._markerMap.get(guid);
-          if (markerData !== undefined ) {
-            selected.push(values[markerData.current]);
-          }
+        for (const marker of this._selected.values()) {
+            selected.push(values[marker.current]);
         }
         return this._selectTrace(values, selected, trace);
     }
@@ -1025,29 +1011,29 @@ export class PropertiesMap {
      * all of them if `trace === undefined`.
      */
     private _symbols(trace?: number): Array<number | number[] | string | string[]> {
-        if (this._settings.symbol.value !== '') {
-            const values = this._property(this._settings.symbol.value).values;
[…]
