### 2020-06-08 ece0c26
Reformat structure.ts
