+++ src/structure/structure.ts
@@ -11,9 +11,8 @@ import {EnvironmentIndexer, generateGUID, getByID, getNextColor, Indexes, sendWa
 import {structure2JSmol} from './jsmol';
 import {JSmolWidget} from './widget';
 
-import close_svg from './close.svg';
-import duplicate_svg from './duplicate.svg';
-import HTML_SETTINGS from './settings.html';
+import CLOSE_SVG from './close.svg';
+import DUPLICATE_SVG from './duplicate.svg';
 
 const MAX_WIDGETS = 9;
 
@@ -21,35 +20,24 @@ const MAX_WIDGETS = 9;
  * Function to return *optimal* arrangement of n widgets.
  * Defaults to a 4 x ? grid. Hardcoded for simplicity.
  */
-function bestArrangement(n: number) {
-  switch (n) {
+function bestGridArrangement(n: number) {
+    switch (n) {
     case 1:
     case 2:
-      return {rows: n, columns: 1};
-      break;
+        return {rows: n, columns: 1};
     case 3:
     case 4:
-      return {rows: 2, columns: 2};
-      break;
+        return {rows: 2, columns: 2};
     case 5:
     case 6:
-      return {rows: 3, columns: 2};
-      break;
+        return {rows: 3, columns: 2};
     case 8:
     case 7:
     case 9:
-      return {rows: 3, columns: 3};
-      break;
-    case 10:
-    case 11:
-    case 12:
-      return {rows: 4, columns: 3};
-      break;
+        return {rows: 3, columns: 3};
     default:
-      return {rows: 4, columns: 4};
-      break;
-
-  }
+        throw Error(`Can not create a grid with more than ${MAX_WIDGETS} elements`);
+    }
 }
 
 /*
@@ -101,10 +89,11 @@ function checkStructure(s: any): string {
 }
 
 interface WidgetGridData {
-        widget: JSmolWidget;
-        color: string;
-        current: Indexes;
-        }
+    widget: JSmolWidget;
+    color: string;
+    current: Indexes;
+}
+
 /**
  * The [[StructureViewer]] class displays a molecule or a crystal in 3D using
  * [JSmol](http://wiki.jmol.org/index.php/JSmol) for rendering.
@@ -205,8 +194,8 @@ export class StructureViewer {
         this.active = this._GUIDs[0];
         this._active = this._GUIDs[0];
         this._delay = this._setDelay();
-
     }
+
     /**
      * Show a new structure, as identified by `indexes`. This will switch to
      * the structure at index `indexes.structure`, and if environments where
@@ -217,37 +206,39 @@ export class StructureViewer {
      * @param  wi              GUID of the widget to show
      */
     public show(indexes: Indexes, selectedGUID: string = this._active) {
-        const widgetData = this._widgetMap.get(selectedGUID);
-        if (widgetData !== undefined) {
-          const widget = widgetData.widget;
-          if (widgetData.current.structure !== indexes.structure) {
-              const options = {
-                  packed: false,
-                  trajectory: true,
-              } as any;
-              assert(indexes.structure < this._structures.length);
+        const data = this._widgetMap.get(selectedGUID);
+        assert(data !== undefined);
 
-              if (this._environments !== undefined) {
-                  options.environments = this._environments[indexes.structure];
-                  if (this._indexer.mode === 'atom') {
-                      options.highlight = indexes.atom;
-                  }
-              }
+        const widget = data.widget;
+        if (data.current.structure !== indexes.structure) {
+            const options = {
+                packed: false,
+                trajectory: true,
+            } as any;
+            assert(indexes.structure < this._structures.length);
 
-              widget.load(`inline '${this._structureForJSmol(indexes.structure)}'`, options);
-          }
[…]
