### 2020-06-08 33cd64d
More refactoring & added documentation
