### 2020-06-09 52c99f5
Removed unnecessary parameter _GUIDs and function _resizeGridInc from src/structure/structure.ts
