+++ src/structure/structure.ts
@@ -128,9 +128,7 @@ export class StructureViewer {
     // GUID of the Active Widget
     private _active: string;
     // Map of Widgets GUIDS to their color, widget, and current structure
-    private _widgetMap: Map<string, WidgetGridData>;
-    // List of Widgets GUIDS in the StructureViewer
-    private _GUIDs: string[];
+    private _selected: Map<string, WidgetGridData>;
     // Documentation needed.
     private _root: HTMLElement;
     /// Playback delay setting
@@ -185,9 +183,7 @@ export class StructureViewer {
         };
 
         // Initializes with 1 widget upon opening.
-        this._widgetMap = new Map();
-        // this._widgetMap = {};
-        this._GUIDs = [];
+        this._selected = new Map();
         this.onselect = () => {};
 
         // Initialize the _root div as a grid (see function below)
@@ -201,8 +197,8 @@ export class StructureViewer {
         root.appendChild(this._root);
 
         this._setupGrid(1);
-        this.active = this._GUIDs[0];
-        this._active = this._GUIDs[0];
+        this.active = this._selected.keys().next().value;
+        this._active = this._selected.keys().next().value;
 
         // get the 'delay' setting inside the current widget setting
         // TODO(pinning): this needs to be updated when another widget is made active
@@ -219,7 +215,7 @@ export class StructureViewer {
      * @param  wi              GUID of the widget to show
      */
     public show(indexes: Indexes, selectedGUID: string = this._active) {
-        const data = this._widgetMap.get(selectedGUID);
+        const data = this._selected.get(selectedGUID);
         assert(data !== undefined);
 
         const widget = data.widget;
@@ -262,11 +258,8 @@ export class StructureViewer {
      * opened.
      */
     public settingsPlacement(callback: (rect: DOMRect) => {top: number, left: number}) {
-        for (const guid of this._GUIDs) {
-            const widgetData = this._widgetMap.get(guid);
-            if (widgetData !== undefined ) {
-                widgetData.widget.settingsPlacement(callback);
-            }
+        for (const widgetData of this._selected.values()) {
+            widgetData.widget.settingsPlacement(callback);
         }
     }
 
@@ -277,7 +270,7 @@ export class StructureViewer {
     public structurePlayback(advance: () => boolean) {
         setTimeout(() => {
             if (advance()) {
-                const widgetData = this._widgetMap.get(this._active);
+                const widgetData = this._selected.get(this._active);
                 if (widgetData !== undefined) {
                   const current = widgetData.current;
                   const structure = (current.structure + 1) % this._indexer.structuresCount();
@@ -298,7 +291,7 @@ export class StructureViewer {
     public atomPlayback(advance: () => boolean) {
         setTimeout(() => {
             if (advance()) {
-                const widgetData = this._widgetMap.get(this._active);
+                const widgetData = this._selected.get(this._active);
                 if (widgetData !== undefined) {
                   const current = widgetData.current;
                   const structure = current.structure;
@@ -317,7 +310,7 @@ export class StructureViewer {
      * Remove all HTML added by this [[StructureViewer]] in the current document
      */
     public remove(): void {
-        for (const data of this._widgetMap.values()) {
+        for (const data of this._selected.values()) {
             data.widget.remove();
         }
         this._root.parentElement!.innerHTML = '';
@@ -358,10 +351,9 @@ export class StructureViewer {
     public set active(activeGUID: string) {
         /// this is here to prevent infinite loops
         if (activeGUID !== this._active) {
-            const newActiveIdx = this._GUIDs.indexOf(activeGUID);
             let indexes;
 
-            if (newActiveIdx >= 0) {
+            if (this._selected.has(activeGUID) ) {
                 if (this._active !== '' && this._active !== undefined) {
[…]
