### 2020-06-09 04315fe
Option to automatically center the selected environment in the viewer (#32)
Also rearranged the environment options to make space for the centering switch
