### 2020-06-09 7f94fb3
Removed instances of this._active === undefined for more thorough checks
