### 2020-06-09 7efc0a3
Fixed 2D-3D marker mappings
