### 2020-06-09 b483b0f
Removed x = map.get(); if (x===undefined)... etc and replaced with assertions/loops where appropriate
