### 2020-06-09 7484686
Added map updateAllMarkers and removed need for checking map instances in updateSelectedMarker
