### 2020-06-09 9022fbd
Made activate, close, and duplicate buttons 'divs' because the 'button' class overrides tooltiptext placement/highlighting, and 'btn/btn-sm' (in case of active) distorts shape
