### 2020-06-09 0130948
Made similar changes in map.ts
