### 2020-06-09 31dae5f
Not very elegant, but at least we avoid an ugly error box that has apparently no reason and no impact
