### 2020-06-09 cd855b6
Resolving todo: playback resets when a new active is selected
