### 2020-06-09 3134106
Resolving todo: guid description listed in widget.ts constructor
