### 2020-06-09 7bc0a71
Resolving todo: Added @param descriptions for _addMarker and _updateMarker
