### 2020-06-09 cfd54fd
Resolving todo: Added @param descriptions for _removeMarker
