### 2020-06-09 61ba3fe
Update src/map/map.ts
Co-authored-by: Guillaume Fraux <luthaf@luthaf.fr>
