### 2020-06-09 5252fba
Added some docs for MarkerData
