### 2020-06-09 4543d6c
Made it possible to instantiate a map without a starterGUID
