+++ src/map/map.ts
@@ -7,7 +7,7 @@ import assert from 'assert';
 
 import {Property} from '../dataset';
 import {EnvironmentIndexer, HTMLSetting, Indexes, SettingGroup, SettingModificationOrigin} from '../utils';
-import {foreachSetting, getByID, makeDraggable, sendWarning} from '../utils';
+import {foreachSetting, getByID, generateGUID, makeDraggable, sendWarning} from '../utils';
 
 import Plotly from './plotly/plotly-scatter';
 import {Config, Data, Layout, PlotlyScatterElement} from './plotly/plotly-scatter';
@@ -245,11 +245,12 @@ export class PropertiesMap {
      * @param indexer    [[EnvironmentIndexer]] used to translate indexes from
      *                   environments index to structure/atom indexes
      * @param properties properties to be displayed
+     * @param starterGUID
      */
     constructor(id: string,
                 indexer: EnvironmentIndexer,
                 properties: {[name: string]: Property},
-                starterGuid: string, /// fix me later
+                starterGUID?: string
                 ) {
         this._indexer = indexer;
         this.onselect = () => {};
@@ -268,8 +269,14 @@ export class PropertiesMap {
         this._root.appendChild(this._plot);
 
         this._data = new MapData(properties);
-        this._addMarker(starterGuid);
-        this.active = starterGuid;
+
+        if (starterGUID !== null) {
+          this._addMarker(starterGUID);
+          this.active = starterGUID;
+        } else {
+          this._addMarker(generateGUID());
+          this.active = this._selected.keys().next().value;
+        }
 
         this._insertSettingsHTML();
         this._colorReset = getByID<HTMLButtonElement>('chsp-color-reset');
