### 2020-06-09 6d07a76
Fixed a quick typo and added a comment on this._active=''
