### 2020-06-09 ab0676d
Update src/structure/structure.ts
Co-authored-by: Guillaume Fraux <luthaf@luthaf.fr>
