### 2020-06-09 637dcda
Tweak styling of tooltips
