### 2020-06-09 cedc132
Remove unused parameter
