### 2020-06-09 22522a4
Cleanup assignment to this._active
It is intialized through this.active setter
