### 2020-06-09 bdd14e5
Cleanup color selection in the viewer grid
