### 2020-06-09 1829955
Small cosmetic changes, more homogeneous style using fa icons
