### 2020-06-10 fd7c64c
Only have a single function for structure checking
