### 2020-06-10 ab35b60
Extract a function to get the next available color
