### 2020-06-10 058be53
Extract a function to get the next available color
