### 2020-06-10 c405a6d
Use template to generate DOM objects from HTML strings
And simplify CSS related to selected structure markers. In particular
even if the marker is not for the currently active one (i.e. is smaller)
it will still stay centered on the corresponding point of the map
