### 2020-06-10 29baab7
Use SVG for the hamburger menus too
