### 2020-06-10 a192176
Use SVG for the hamburger menus too
