### 2020-06-11 58114b2
Decouple grid creation and showing structures in viewers
This allow "on demand loading" of structures
