+++ src/structure/structure.ts
@@ -6,7 +6,8 @@
 import assert from 'assert';
 
 import {Environment, isPositiveInteger, Structure, UserStructure} from '../dataset';
-import {EnvironmentIndexer, generateGUID, getByID, getNextColor, Indexes, sendWarning} from '../utils';
+import {EnvironmentIndexer, Indexes, PositioningCallback} from '../utils';
+import {generateGUID, getByID, getNextColor, sendWarning} from '../utils';
 
 import {structure2JSmol} from './jsmol';
 import {JSmolWidget} from './widget';
@@ -144,7 +145,9 @@ export class StructureViewer {
     /// GUID of the Active Widget
     private _active: string;
     /// Map of Widgets GUIDS to their color, widget, and current structure
-    private _selected: Map<string, WidgetGridData>;
+    private _viewers: Map<string, WidgetGridData>;
+    /// Callback used to override all grid viewers' positionSettingsModal
+    private _positionSettingsModal?: PositioningCallback;
 
     /**
      * Create a new [[StructureViewer]] inside the HTML element with the given
@@ -184,22 +187,18 @@ export class StructureViewer {
         };
 
         // Initializes with 1 widget upon opening.
-        this._selected = new Map();
+        this._viewers = new Map();
         this.onselect = () => {};
 
-        // Initialize the _root div as a grid (see function below)
         const root = getByID(id);
-
-        // Initialize the grid root
         this._root = document.createElement('div');
         this._root.id = 'grid-root';
         this._root.className = 'chsp-structure-viewer-grid';
-
         root.appendChild(this._root);
 
         this._setupGrid(1);
-        this.active = this._selected.keys().next().value;
-        this._active = this._selected.keys().next().value;
+        this.active = this._viewers.keys().next().value;
+        this._active = this._viewers.keys().next().value;
 
         // get the 'delay' setting inside the current widget setting
         this._delay = getByID<HTMLInputElement>(`chsp-${this._active}-playback-delay`);
@@ -215,7 +214,7 @@ export class StructureViewer {
      * @param  selectedGUID   GUID of the widget to show
      */
     public show(indexes: Indexes, selectedGUID: string = this._active) {
-        const data = this._selected.get(selectedGUID);
+        const data = this._viewers.get(selectedGUID);
         assert(data !== undefined);
 
         const widget = data.widget;
@@ -249,28 +248,13 @@ export class StructureViewer {
     }
 
     /**
-     * Register a `callback` to compute the placement of the settings modal.
-     *
-     * The callback gets the current placement of the settings as a
-     * [DOMRect](https://developer.mozilla.org/en-US/docs/Web/API/DOMRect),
-     * and should return top and left positions in pixels, used with `position:
-     * fixed`. The callback is called once, the first time the settings are
-     * opened.
-     */
-    public settingsPlacement(callback: (rect: DOMRect) => {top: number, left: number}) {
-        for (const widgetData of this._selected.values()) {
-            widgetData.widget.settingsPlacement(callback);
-        }
-    }
-
-    /**
      * Start playing the trajectory of structures in this dataset, until
      * `advance` returns false
      */
     public structurePlayback(advance: () => boolean) {
         setTimeout(() => {
             if (advance()) {
-                const widgetData = this._selected.get(this._active);
+                const widgetData = this._viewers.get(this._active);
                 assert(widgetData !== undefined);
 
                 const current = widgetData.current;
@@ -292,7 +276,7 @@ export class StructureViewer {
     public atomPlayback(advance: () => boolean) {
         setTimeout(() => {
             if (advance()) {
-                const widgetData = this._selected.get(this._active);
+                const widgetData = this._viewers.get(this._active);
                 assert(widgetData !== undefined);
 
[…]
