### 2020-06-11 b037761
Make sure the modal positionning is the same for all viewers
And unify interface by directly setting the callback as a property
of the PropertyMap/StructureViewerGrid/JSmolWidget
