### 2020-06-11 d77f7dc
Simplify a loop
