### 2020-06-11 014dc67
Make dataset removal a bit more robust
