### 2020-06-11 21ac1be
Synchronized 3D marker color with structure viewer
