### 2020-06-11 b753fc9
Updating markers that have just been added
