### 2020-06-11 1056606
Synchronizing marker sizes with active/inactive
