### 2020-06-11 665fd62
Typo in checking 3D
