### 2020-06-11 2fae89c
Fix on demand loading of structure
The code no longer tries to load a structure in the constructor
