### 2020-06-11 d383f99
Fix on demand loading of structure
The code no longer tries to load a structure in the constructor
