+++ src/utils/index.ts
@@ -33,7 +33,7 @@ export function getNextColor(alreadyUsedColors: string[]) {
     throw Error('required more colors than available');
 }
 
-function generateGUID() {
+function generateGUID(): string {
     return 'xxxxxxxx-xxxx-4xxx-yxxx-xxxxxxxxxxxx'.replace(/[xy]/g, (c) => {
         // tslint:disable-next-line:no-bitwise
         const r = Math.random() * 16 | 0;
