### 2020-06-11 9f093e1
Remove unused variable
