### 2020-06-11 02bed8b
added a flag to the selected trace
