### 2020-06-11 1d928eb
Made the selected markers in 3D responsible to clicking re:active/inactive. Had to do it somewhat hackily because the only properties available in the click event for comparison require iteration through all selected points. Will become slow if there are a lot of selected points.
