### 2020-06-11 3829ac3
Fixes switch-active-on-2D-redraw bug
