### 2020-06-11 c05661e
I put together a thorough fix of the active systems. This included rewriting the set active function within structure.ts, which had been rewritten to streamline the code, but missed a few nuances:
	- structure.ts:
		- whenever you set the active button, it puts a whole domino effect in motion. Therefore we should avoid doing `this._active = guid` as it will ignore this domino effect.
		- when the grid is switching datasets/setting up anew/etc. there will be no `old` active
		- if the grid tries to select an active which doesn't exist on the grid, do nothing
		- there is now a catch to check if
[…]
