### 2020-06-12 d701781
Enable source maps in development
