### 2020-06-12 fdd9233
Don't try to update the new widget if it doesn't exists
