### 2020-06-12 fd13fc6
Minor tweaks & formatting
