+++ src/map/map.ts
@@ -832,7 +832,7 @@ export class PropertiesMap {
         // only used in 3D mode, since it is way slower than moving
         // this._selectedMarker around.
         const selected = {
-            name: '',
+            name: 'selected',
             type: 'scattergl',
 
             x: [NaN],
@@ -844,8 +844,9 @@ export class PropertiesMap {
                 color: colors[1],
                 line: {
                     color: lineColors[1],
-                    width: 1,
+                    width: 2,
                 },
+                opacity: 1,
                 size: sizes[1],
                 sizemode: 'area',
             },
@@ -904,28 +905,26 @@ export class PropertiesMap {
                 return;
             }
 
-            let indexes; // container for the indices
-            let md; // container for the marker data
-
+            let data;
             // if someone has clicked on a selection marker, set to active
             if (event.points[0].data.name === 'selected' ) {
-              let i = 0;
-              for (const [guid, markerData] of this._selected.entries()) {
-                if (i === event.points[0].pointNumber) {
-                    this.active = guid;
-                    md = markerData;
-                    break;
+                let i = 0;
+                for (const [guid, markerData] of this._selected.entries()) {
+                    if (i === event.points[0].pointNumber) {
+                        this.active = guid;
+                        data = markerData;
+                        break;
                     }
-                i++;
-              }
+                    i++;
+                }
             }
-            if (md !== undefined) {
-              indexes = this._indexer.from_environment(md.current);
-            } else {
-              // default if someone has clicked generally on the map or on a
-              // place the selected marker doesn't recognize
-              indexes = this._indexer.from_environment(event.points[0].pointNumber);
-            }
+
+            const indexes = (data !== undefined) ?
+                this._indexer.from_environment(data.current) :
+                // default if someone has clicked generally on the map or on a
+                // place the selected marker doesn't recognize
+                this._indexer.from_environment(event.points[0].pointNumber);
+
             this.select(indexes, this._active);
             this.onselect(indexes, this._active);
         });
@@ -966,8 +965,10 @@ export class PropertiesMap {
         const selected = [];
         for (const marker of this._selected.values()) {
             if (this._is3D()) {
-              selected.push(values[marker.current]);
-            } else { selected.push(NaN); }
+                selected.push(values[marker.current]);
+            } else {
+                selected.push(NaN);
+            }
         }
         return this._selectTrace(values, selected, trace);
     }
@@ -981,8 +982,10 @@ export class PropertiesMap {
         const selected = [];
         for (const marker of this._selected.values()) {
             if (this._is3D()) {
-              selected.push(values[marker.current]);
-            } else { selected.push(NaN); }
+                selected.push(values[marker.current]);
+            } else {
+                selected.push(NaN);
+            }
         }
 
         return this._selectTrace(values, selected, trace);
@@ -1009,7 +1012,7 @@ export class PropertiesMap {
      * Get the color values to use with the given plotly `trace`, or all of
      * them if `trace === undefined`
      */
-    private _colors(trace?: number): Array<string | number | number[]> {
+    private _colors(trace?: number): Array<string | string[] | number | number[]> {
         let values;
         if (this._hasColors()) {
             values = this._property(this._settings.color.property.value).values;
@@ -1017,7 +1020,12 @@ export class PropertiesMap {
             values = 0.5;
         }
 
-        return this._selectTrace<string | number | number[]>(values, '#007bff', trace);
+        const selected = [];
+        for (const data of this._selected.values()) {
+            selected.push(data.color);
+        }
+
[…]
