### 2020-06-12 94395c2
Fixes a bug with the standalone generator.
