### 2020-06-13 fd4ccd3
Better logic for including a loading box in standalone.html
