### 2020-06-14 6344d13
Intial implementation of structure viewser settings
