### 2020-06-14 3e4b51a
Intial implementation of structure viewser settings
