+++ src/structure/settings.ts
@@ -1,0 +1,164 @@
+/**
+ * @packageDocumentation
+ * @module settings
+ */
+
+import {HTMLSetting} from '../utils';
+
+interface EnvironmentPresets {
+    activated: boolean;
+    center: boolean;
+    cutoff: number;
+    bgStyle: string;
+    bgColor: string;
+}
+
+export interface StructurePresets {
+    bonds: boolean;
+    spaceFilling: boolean;
+    atomLabels: boolean;
+    unitCell: boolean;
+    packedCell: boolean;
+    supercell: number[];
+    rotation: boolean;
+    axes: string;
+    environments: Partial<EnvironmentPresets>;
+    keepOrientation: boolean;
+}
+
+export const STRUCTURE_DEFAULTS: StructurePresets = {
+    atomLabels: false,
+    axes: 'off',
+    bonds: true,
+    environments: {
+        activated: true,
+        bgColor: 'grey',
+        bgStyle: 'licorice',
+        center: false,
+        cutoff: 4.0,
+    },
+    keepOrientation: false,
+    packedCell: false,
+    rotation: false,
+    spaceFilling: false,
+    supercell: [1, 1, 1],
+    unitCell: false,
+};
+
+export class StructureSettings {
+    // should we show bonds
+    public bonds: HTMLSetting<'boolean'>;
+    // should we use space filling representation
+    public spaceFilling: HTMLSetting<'boolean'>;
+    // should we show atoms labels
+    public atomLabels: HTMLSetting<'boolean'>;
+    // should we show unit cell information and lines
+    public unitCell: HTMLSetting<'boolean'>;
+    /// Is the current unit cell displayed as a packed cell?
+    public packedCell: HTMLSetting<'boolean'>;
+    /// number of repetitions in the `a/b/c` direction for the supercell
+    public supercell: [HTMLSetting<'int'>, HTMLSetting<'int'>, HTMLSetting<'int'>];
+    // should we spin the represenation
+    public rotation: HTMLSetting<'boolean'>;
+    // which axis system to use (none, xyz, abc)
+    public axes: HTMLSetting<'string'>;
+    // keep the orientation constant when loading a new structure if checked
+    public keepOrientation: HTMLSetting<'boolean'>;
+    // options related to environments
+    public environments: {
+        // should we display environments & environments options
+        activated: HTMLSetting<'boolean'>;
+        // automatically center the environment when loading it
+        center: HTMLSetting<'boolean'>;
+        // the cutoff value for spherical environments
+        cutoff: HTMLSetting<'number'>;
+        // which style for atoms not in the environment
+        bgStyle: HTMLSetting<'string'>;
+        // which colors for atoms not in the environment
+        bgColor: HTMLSetting<'string'>;
+    };
+
+    constructor(presets: Partial<StructurePresets> = {}) {
+        this.bonds = new HTMLSetting('boolean', STRUCTURE_DEFAULTS.bonds);
+        this.spaceFilling = new HTMLSetting('boolean', STRUCTURE_DEFAULTS.spaceFilling);
+        this.atomLabels = new HTMLSetting('boolean', STRUCTURE_DEFAULTS.atomLabels);
+        this.unitCell = new HTMLSetting('boolean', STRUCTURE_DEFAULTS.unitCell);
+        this.packedCell = new HTMLSetting('boolean', STRUCTURE_DEFAULTS.packedCell);
+        this.supercell = [
+            new HTMLSetting('int', STRUCTURE_DEFAULTS.supercell[0]),
+            new HTMLSetting('int', STRUCTURE_DEFAULTS.supercell[1]),
[…]
