### 2020-06-14 9461bfe
Intial implementation of structure viewser settings
