### 2020-06-15 0fe9e8d
Remove short option for loader
