+++ utils/generate_standalone.py
@@ -75,9 +75,9 @@ def inline_ressources(html):
 def main():
     """
     Command-line utility to generate an stand-alone HTML file containing the
-    chemiscope app. The standalone HTML can be loaded in a browser, and used to 
+    chemiscope app. The standalone HTML can be loaded in a browser, and used to
     read chemiscope JSON inputs. In addition, a chemiscope JSON can be appended
-    to the HTML to have a fully self-contained viewer for a chosen data set. 
+    to the HTML to have a fully self-contained viewer for a chosen data set.
     """
 
     # Tweak the autogenerated help output to look nicer
@@ -87,10 +87,10 @@ def main():
                         help='chemiscope JSON input that is appended to provide default data')
     parser.add_argument('-o', '--output', type=str, default="",
                         help='output file name. Defaults to chemiscope.html or the name of the JSON')
-    parser.add_argument('-l', '--loader', default=False, action='store_true',
+    parser.add_argument('--loader', default=False, action='store_true',
                         help='shows a load input box that makes it possible to load new files even if there is an embedded dataset.')
     args = parser.parse_args()
-    
+
     html = get_html()
     inline_ressources(html)
 
@@ -119,9 +119,8 @@ def main():
             fd.write('<script type="text/javascript">  document.getElementById("upload-form").style.display = "none";  </script>')
         fd.write('<script id=standalone-json-data type="application/json">')
         fd.write(json_data)
-        
 
-    print('created '+ output_name +". happy chemiscoping!")
+    print('created ' + output_name + ". Happy chemiscoping!")
 
 
 if __name__ == '__main__':
