### 2020-06-15 554d52c
Fixes a bug with the standalone generator. (#46)
