+++ utils/generate_standalone.py
@@ -75,9 +75,9 @@ def inline_ressources(html):
 def main():
     """
     Command-line utility to generate an stand-alone HTML file containing the
-    chemiscope app. The standalone HTML can be loaded in a browser, and used to 
+    chemiscope app. The standalone HTML can be loaded in a browser, and used to
     read chemiscope JSON inputs. In addition, a chemiscope JSON can be appended
-    to the HTML to have a fully self-contained viewer for a chosen data set. 
+    to the HTML to have a fully self-contained viewer for a chosen data set.
     """
 
     # Tweak the autogenerated help output to look nicer
@@ -87,8 +87,10 @@ def main():
                         help='chemiscope JSON input that is appended to provide default data')
     parser.add_argument('-o', '--output', type=str, default="",
                         help='output file name. Defaults to chemiscope.html or the name of the JSON')
+    parser.add_argument('--loader', default=False, action='store_true',
+                        help='shows a load input box that makes it possible to load new files even if there is an embedded dataset.')
     args = parser.parse_args()
-    
+
     html = get_html()
     inline_ressources(html)
 
@@ -99,21 +101,26 @@ def main():
         else:
             output_name = os.path.splitext(args.input)[0] + ".html"
 
+    # show loader by default
+    show_loader = True
     json_data = ""
     if args.input != "":
         if args.input.endswith(".gz"):
             raise ValueError("Only plain-text JSON inputs can be used with the standalone viewer")
         with open(args.input) as fd:
             json_data = fd.read()
+        # if an input json is provided, check the -l option (wich defaults to False)
+        show_loader = args.loader
 
     with open(output_name, 'w') as fd:
         fd.write(str(html))
         fd.write('<script type="text/javascript"> for (const element of document.getElementsByClassName("hide-if-standalone") ) { element.style.display = "none"; } </script>')
+        if not show_loader:
+            fd.write('<script type="text/javascript">  document.getElementById("upload-form").style.display = "none";  </script>')
         fd.write('<script id=standalone-json-data type="application/json">')
         fd.write(json_data)
-        
 
-    print('created '+ output_name +". happy chemiscoping!")
+    print('created ' + output_name + ". Happy chemiscoping!")
 
 
 if __name__ == '__main__':
