+++ .travis.yml
@@ -1,56 +1,0 @@
-dist: bionic
-
-language: node_js
-addons:
-  apt:
-    packages:
-      - python3
-      - python3-setuptools
-      - python3-pip
-
-node_js:
-  - 12
-
-install:
-    # we need the full history to be able to get the chemiscope version with
-    # git describe
-    - git fetch --unshallow
-    - npm ci
-    - python3 -m pip install -r docs/requirements.txt    
-    - python3 -m pip install beautifulsoup4 requests
-
-script:
-    # Build the code
-    - npm run build
-    - npm run merge-dts
-    # Run tests & lints
-    - npm test
-    # Build the website & standalone version
-    - npm run download-example-input
-    - cp -r app/ gh-pages/
-    - cp dist/*.min.js gh-pages/    
-    - python3 utils/generate_standalone.py -o gh-pages/chemiscope_standalone.html    
-    - touch gh-pages/.nojekyll
-    - echo "chemiscope.org" > gh-pages/CNAME
-    # Cleanup a few files
-    - rm -f gh-pages/.gitignore
-    # Build the docs
-    - cd docs && make html && cd -
-    - cp -r docs/_build/html/ gh-pages/docs/
-
-deploy:
-  # Deploy the website to gh-pages
-  - provider: pages
-    skip_cleanup: true
-    github_token: $GH_TOKEN
-    local_dir: gh-pages
-  # Deploy built code to github releases
-  - provider: releases
-    api_key: $GH_TOKEN
-    file:
-      - "dist/chemiscope.d.ts"
-      - "dist/chemiscope.min.js"
-      - "dist/jsmol-widget.min.js"
-    skip_cleanup: true
-    on:
-      tags: true
