### 2020-06-15 074abb8
Also deploy the standalone chemiscope
