### 2020-06-15 bceac20
Also upload minified JS to releases
