### 2020-06-15 fa13d92
More deploy config tweaks
