### 2020-06-15 cosmo-epfl/chemiscope#47 @kblin
Failed to run `npm run download-example-input` when /tmp is on a different partition than chemiscope
When trying to download the example input following the install instructions, the `download-example-input` step fails when `/tmp` is on a different partition than the current working directory:

```
$ npm run download-example-input

> chemiscope@0.1.0 download-example-input /home/kai/Documents/reviews/2020/JOSS-2117/chemiscope
> ts-node ./utils/download-example-input.ts

Cloning into '/tmp/tmp-173153-8SqGyGFytJyu/chemiscope'...
Error: EXDEV: cross-device link not permitted, rename '/tmp/tmp-173153-8SqGyGFytJyu/chemiscope/CSD-500.json.gz' -> './app/CSD-500.json.gz'
    at Object.renameSync (fs.js:756:3)
    at Object.<anonymous> (/home/kai/Documents/reviews/2020/JOSS-2117/chemiscope/utils/download-example-input.ts:14:8)
    at Module._compile (internal/modules/cjs/loader.js:1200:30)
    at Module.m._compile (/home/kai/Documents/reviews/2020/JOSS-2117/chemiscope/node_modules/ts-node/src/index.ts:858:23)
    at Module._extensions..js (internal/modules/cjs/loader.js:1220:10)
    at Object.require.extensions.<computed> [as .ts] (/home/kai/Documents/reviews/2020/JOSS-2117/chemiscope/node_modules/ts-node/src/index.ts:861:12)
    at Module.load (internal/modules/cjs/loader.js:1049:32)
    at Function.Module._load (internal/modules/cjs/loader.js:937:14)
    at Function.executeUserEntryPoint [as runMain] (internal/modules/run_main.js:71:12)
    at main (/home/kai/Documents/reviews/2020/JOSS-2117/chemiscope/node_modules/ts-node/src/bin.ts:227:14)
```
Googling the error message ends me up on a [stackoverflow post](https://stackoverflow.com/questions/43206198/what-does-the-exdev-cross-device-link-not-permitted-error-mean) that sounds like `fs.rename` uses a syscall that doesn't work across partitions/mount points/filesystems.
In my case, `/home` is on a different filesystem than `/tmp`, and thus the above error happens.
