### 2020-06-15 cosmo-epfl/chemiscope#47 @kblin
I forgot to mention that cloning chemiscope in `/tmp` of course makes this work fine.
