### 2020-06-15 #2117 @kblin
I've found two minor issues during the review:
* When installing into a folder that that's not on the same filesystem as `/tmp`, downloading the example data fails. I've filed a bug: https://github.com/cosmo-epfl/chemiscope/issues/47
* Manuscript page 2, 2nd paragraph. Citation for Bartók, 2010 isn't resolved properly
