### 2020-06-15 cosmo-epfl/chemiscope#47 @Luthaf
That's an interesting one, thanks for the report!

We may as well do the download directly inside the chemiscope folder then.
