### 2020-06-16 cosmo-epfl/chemiscope#47 @kblin
Yeah, I was a bit surprised that `fs.rename` apparently tries to hardlink the file to the new location before deleting the old copy. It should be smart enough to fall back to some other strategy if that doesn't work. But I agree downloading directly in the chemiscope folder would also fix things.
