### 2020-06-16 866dc6b
Fix download-example-input if /tmp is on a different partition
