### 2020-06-16 e55ee6e
Fix reference in JOSS paper
