### 2020-06-16 cosmo-epfl/chemiscope#47 @Luthaf
Should be fixed now, I would appreciate if you could test it since I don't have a system where `/tmp` is on a different partition.
