### 2020-06-16 cosmo-epfl/chemiscope#47 @kblin
Yes, fixed now, thanks!
