### 2020-06-19 18f8a81
Small edits to the paper
