+++ paper/paper.md
@@ -22,9 +22,9 @@ bibliography: paper.bib
 
 # Summary
 
-The number of different materials or molecules that can created by combining
+The number of materials or molecules that can be created by combining
 different chemical elements in various proportions and spatial arrangements is
-enormous. We can use computational chemistry to generate databases containing
+enormous. Computational chemistry can be used to generate databases containing
 billions of potential structures [@Ruddigkeit2012], and predict some of the
 associated properties [@Montavon2013; @Ramakrishnan2014]. Unfortunately, the
 very large number of structures makes exploring such database — to understand
@@ -57,11 +57,11 @@ descriptors can be very high dimensional, it can be convenient to apply a
 dimensionality reduction algorithm that maps them to a lower-dimensional space
 for easier visualization. For example the sketchmap algorithm [@Ceriotti2011]
 was used with the Smooth Overlap of Atomic Positions descriptor [@Bartok2013] to
-generate the visualization above. The right panel displays the three-dimensional
-structure of the chemical entities, possibly including periodic repetition for
-crystals. Visualizing the chemical structure can help finding an intuitive
-rationalization of the layout of the dataset and the structure-property
-relations.
+generate the visualization figure 1. The right panel displays the
+three-dimensional structure of the chemical entities, possibly including
+periodic repetition for crystals. Visualizing the chemical structure can help
+finding an intuitive rationalization of the layout of the dataset and the
+structure-property relations.
 
 Whereas similar tools [@Gong2013; @Gutlein2014; @Probst2017; @ISV] only allow
 visualizing maps and structures in which each data point corresponds to a
@@ -74,7 +74,7 @@ machine-learning schemes that decompose properties into atom-centred
 contributions [@Behler2007; @Bartok2010]
 
 
-![Database of chemical shielding [@Paruzzo2018] in chemiscope showing the use of a 3D plot and atomic environments highlighting](./screenshot-3d.png)
+![Database of chemical shieldings [@Paruzzo2018] in chemiscope demonstrating the use of a 3D plot and atomic environments highlighting](./screenshot-3d.png)
 
 Chemiscope took strong inspiration from a previous similar graphical software,
 the interactive sketchmap visualizer [@ISV]. This previous software was used in
@@ -90,20 +90,20 @@ interactivity. It uses [Plotly.js](https://plot.ly/javascript/) to render and
 animate 2D and 3D plots; and the JavaScript version of [Jmol](http://jmol.org/)
 to display atomic structures. The visualization is fast enough to be used with
 datasets containing up to a million points, reacting to user input within a few
-hundred milliseconds in the default 2D mode. More elaborate visualization are
+hundred milliseconds in the default 2D mode. More elaborate visualizations are
 slower, while still handling 100k points easily.
 
 The use of web technologies makes chemiscope usable from different operating
 systems without the need to develop, maintain and package the code for each
-operating system. It also mean that we can provide an online service at
+operating system. It also means that we can provide an online service at
 http://chemiscope.org allowing users to visualize their own dataset without
-installing anything. Chemiscope is implemented as library of re-usable
+any local installation. Chemiscope is implemented as library of re-usable
 components linked together via callbacks. This makes it easy to modify the
 default interface to generate more elaborate visualizations: for example
 displaying multiple maps generated with different parameters of a dimensionality
 reduction algorithm. Chemiscope can also be distributed in a standalone mode,
 where the code and a predefined dataset are merged together as a single HTML
-file. This standalone mode is useful for archival purposes, for example in as
+file. This standalone mode is useful for archival purposes, for example as
 supplementary information for a published article; and for use in corporate
 environments with sensitive datasets.
 
