### 2020-06-19 f69e71b
Make a proper deep copy of DEFAULT_LAYOUT
Previous code was only making copy at the first level, so changes
to the layout were kept when loading new datasets
