### 2020-06-19 8028e3f
First pass at onremove()
