### 2020-06-22 845061b
Move the code dealing with the DOM/HTML side of settings to settings.ts
