### 2020-06-22 107c952
Move the code dealing with the DOM/HTML side of settings to settings.ts
