+++ src/map/settings.ts
@@ -5,7 +5,14 @@
 
 import assert from 'assert';
 
-import {HTMLSetting, settingsValidator} from '../utils';
+import {HTMLSetting, PositioningCallback} from '../utils';
+import {getByID, makeDraggable, settingsValidator} from '../utils';
+import {NumericProperties} from './data';
+
+import {COLOR_MAPS} from './colorscales';
+
+import BARS_SVG from '../static/bars.svg';
+import HTML_SETTINGS from './settings.html';
 
 export interface AxisPresets {
     /// Which property should we use for this axis
@@ -123,44 +130,61 @@ export class MapSettings {
         factor: HTMLSetting<'number'>;
     };
 
+    /// Default properties used
     private readonly _defaultProperties: Partial<MapPresets>;
+    /// The HTML element containing the settings modal
+    private _settingsModal: HTMLElement;
+    // Callback to get the initial positioning of the settings modal.
+    private _positionSettingsModal: PositioningCallback;
 
-    constructor(properties: string[], presets: Partial<MapPresets> = {}) {
-        if (properties.length < 2) {
+    constructor(
+        root: HTMLElement,
+        properties: NumericProperties,
+        positionSettings: PositioningCallback,
+        presets: Partial<MapPresets> = {},
+    ) {
+        const propertiesName = Object.keys(properties);
+        if (propertiesName.length < 2) {
             throw Error('we need at least two properties to plot in the map');
         }
+
         this._defaultProperties = {
-            x: {property: properties[0]},
-            y: {property: properties[1]},
+            x: {property: propertiesName[0]},
+            y: {property: propertiesName[1]},
         };
 
-        if (properties.length >= 2) {
-            this._defaultProperties.color = {property: properties[2]};
+        if (propertiesName.length >= 2) {
+            this._defaultProperties.color = {property: propertiesName[2]};
         }
 
-        this.x = new AxisSettings(properties);
-        this.y = new AxisSettings(properties);
+        this.x = new AxisSettings(propertiesName);
+        this.y = new AxisSettings(propertiesName);
         // For z and color, '' is a valid value
-        this.z = new AxisSettings(properties.concat(['']));
-        this.color = new AxisSettings(properties.concat(['']));
+        this.z = new AxisSettings(propertiesName.concat(['']));
+        this.color = new AxisSettings(propertiesName.concat(['']));
 
         this.symbol = new HTMLSetting('string', '');
-        this.symbol.validate = settingsValidator(properties.concat(['']), 'symbol');
+        this.symbol.validate = settingsValidator(propertiesName.concat(['']), 'symbol');
 
         this.palette = new HTMLSetting('string', '');
-        this.symbol.validate = settingsValidator(properties.concat(['']), 'symbol');
+        this.palette.validate = settingsValidator(Object.keys(COLOR_MAPS), 'palette');
 
         this.size = {
             factor : new HTMLSetting('number', 0),
             property : new HTMLSetting('string', ''),
         };
-        this.size.property.validate = settingsValidator(properties.concat(['']), 'size');
+        this.size.property.validate = settingsValidator(propertiesName.concat(['']), 'size');
         this.size.factor.validate = (value) => {
             if (value < 0 || value > 100) {
                 throw Error(`size factor must be between 0 and 100, got ${value}`);
             }
         };
 
+        this._positionSettingsModal = positionSettings;
+        this._settingsModal = this._insertSettingsHTML(root);
+        document.body.appendChild(this._settingsModal);
+
+        this._bind(properties);
         this.applyPresets(presets);
     }
 
@@ -205,5 +229,149 @@ export class MapSettings {
             },
             symbol: this.symbol.value,
         };
+    }
+
+    /**
[…]
