+++ src/map/map.ts
@@ -9,16 +9,13 @@ import Plotly from './plotly/plotly-scatter';
 import {Config, Data, Layout, PlotlyScatterElement} from './plotly/plotly-scatter';
 
 import {Property} from '../dataset';
-import {EnvironmentIndexer, Indexes, PositioningCallback, SettingGroup, SettingModificationOrigin} from '../utils';
-import {foreachSetting, generateGUID, getByID, makeDraggable, sendWarning} from '../utils';
+import {EnvironmentIndexer, Indexes, PositioningCallback, SettingModificationOrigin} from '../utils';
+import {generateGUID, getByID, sendWarning} from '../utils';
 
 import {MapData, NumericProperty} from './data';
 import {AxisSettings, MapPresets, MapSettings} from './settings';
 
 import {COLOR_MAPS} from './colorscales';
-
-import BARS_SVG from '../static/bars.svg';
-import HTML_SETTINGS from './settings.html';
 
 const DEFAULT_LAYOUT = {
     // coloraxis is used for the markers
@@ -182,12 +179,10 @@ export class PropertiesMap {
 
     /// environment indexer
     private _indexer: EnvironmentIndexer;
-    /// Store the HTML elements used for settings
-    private _settings!: MapSettings;
+    /// Settings of the map
+    private _settings: MapSettings;
     /// Button used to reset the range of color axis
     private _colorReset: HTMLButtonElement;
-    /// The HTML element containing the settings modal
-    private _settingsModal!: HTMLElement;
 
     /**
      * Create a new [[PropertiesMap]] inside the DOM element with the given HTML
@@ -200,9 +195,9 @@ export class PropertiesMap {
      * @param starterGUID if synchronized with a StructureViewer, the GUID of the first structure viewer cell
      */
     constructor(
-        config: { id: string, presets?: MapPresets },
+        config: { id: string, presets: MapPresets },
         indexer: EnvironmentIndexer,
-        properties: {[name: string]: Property},
+        properties: { [name: string]: Property },
         starterGUID?: string,
     ) {
         this._indexer = indexer;
@@ -229,13 +224,15 @@ export class PropertiesMap {
             this.active = this._selected.keys().next().value;
         }
 
-        this._insertSettingsHTML();
+        this._settings = new MapSettings(
+            this._root,
+            this._data[this._indexer.mode],
+            (rect) => this.positionSettingsModal(rect),
+            config.presets,
+        );
         this._colorReset = getByID<HTMLButtonElement>('chsp-color-reset');
 
-        this._setupSettings(config.presets);
         this._connectSettings();
-
-        this._createPlot();
 
         // By default, position the modal for settings on top of the plot,
         // centered horizontally
@@ -246,6 +243,8 @@ export class PropertiesMap {
                 top: rootRect.top + 20,
             };
         };
+
+        this._createPlot();
     }
 
     /** Change the selected environment to the one with the given `indexes` */
@@ -279,7 +278,6 @@ export class PropertiesMap {
                 markerData.current = indexes.environment;
                 this._updateSelectedMarker(selectedGUID, markerData);
             }
-
         }
     }
 
@@ -288,7 +286,7 @@ export class PropertiesMap {
      */
     public remove(): void {
         this._root.innerHTML = '';
-        this._settingsModal.remove();
+        this._settings.remove();
     }
 
     /**
@@ -337,7 +335,7 @@ export class PropertiesMap {
     /**
      * Apply saved settings to the map.
      */
-    public applyPresets(presets?: MapPresets) {
+    public applyPresets(presets: Partial<MapPresets>) {
         this._settings.applyPresets(presets);
     }
 
[…]
