### 2020-06-22 748029b
Move the code dealing with the DOM/HTML side of settings to settings.ts
