+++ src/map/map.ts
@@ -9,18 +9,15 @@ import Plotly from './plotly/plotly-scatter';
 import {Config, Data, Layout, PlotlyScatterElement} from './plotly/plotly-scatter';
 
 import {Property} from '../dataset';
+
 import {EnvironmentIndexer, Indexes} from '../utils';
-import {SettingGroup, SettingModificationOrigin} from '../utils';
-import {GUID, PositioningCallback} from '../utils';
-import {enumerate, foreachSetting, getByID, getFirstKey, makeDraggable, sendWarning} from '../utils';
+import {GUID, PositioningCallback, SettingModificationOrigin} from '../utils';
+import {enumerate, getByID, getFirstKey, sendWarning} from '../utils';
 
 import {MapData, NumericProperty} from './data';
 import {AxisSettings, MapPresets, MapSettings} from './settings';
 
 import {COLOR_MAPS} from './colorscales';
-
-import BARS_SVG from '../static/bars.svg';
-import HTML_SETTINGS from './settings.html';
 
 const DEFAULT_LAYOUT = {
     // coloraxis is used for the markers
@@ -191,12 +188,10 @@ export class PropertiesMap {
 
     /// environment indexer
     private _indexer: EnvironmentIndexer;
-    /// Store the HTML elements used for settings
-    private _settings!: MapSettings;
+    /// Settings of the map
+    private _settings: MapSettings;
     /// Button used to reset the range of color axis
     private _colorReset: HTMLButtonElement;
-    /// The HTML element containing the settings modal
-    private _settingsModal!: HTMLElement;
 
     /**
      * Create a new [[PropertiesMap]] inside the DOM element with the given HTML
@@ -208,9 +203,9 @@ export class PropertiesMap {
      * @param properties properties to be displayed
      */
     constructor(
-        config: { id: string, presets?: MapPresets },
+        config: { id: string, presets: MapPresets },
         indexer: EnvironmentIndexer,
-        properties: {[name: string]: Property},
+        properties: { [name: string]: Property },
     ) {
         this._indexer = indexer;
         this.onselect = () => {};
@@ -229,13 +224,15 @@ export class PropertiesMap {
 
         this._data = new MapData(properties);
 
-        this._insertSettingsHTML();
+        this._settings = new MapSettings(
+            this._root,
+            this._data[this._indexer.mode],
+            (rect) => this.positionSettingsModal(rect),
+            config.presets,
+        );
         this._colorReset = getByID<HTMLButtonElement>('chsp-color-reset');
 
-        this._setupSettings(config.presets);
         this._connectSettings();
-
-        this._createPlot();
 
         // By default, position the modal for settings on top of the plot,
         // centered horizontally
@@ -246,15 +243,17 @@ export class PropertiesMap {
                 top: rootRect.top + 20,
             };
         };
+
+        this._createPlot();
     }
 
     /**
      * Remove all HTML added by this [[PropertiesMap]] in the current document
      */
-    public remove(): void {
-        this._root.innerHTML = '';
-        this._settingsModal.remove();
-    }
+     public remove(): void {
+         this._root.innerHTML = '';
+         this._settings.remove();
+     }
 
     /**
      * Change the environment indicated by the currently active marker to
@@ -375,7 +374,7 @@ export class PropertiesMap {
     /**
      * Apply saved settings to the map.
      */
-    public applyPresets(presets?: MapPresets) {
+    public applyPresets(presets: Partial<MapPresets>) {
         this._settings.applyPresets(presets);
     }
 
@@ -394,56 +393,6 @@ export class PropertiesMap {
     /** Forward to Ploty.relayout */
     private _relayout(layout: Partial<Layout>) {
[…]
