### 2020-06-22 7690ffb
Removed dependence of onselect on a supplied GUID and removed where no longer applicable
