### 2020-06-22 7aa7a16
Added an 'activate' communicating function between map and structure. set active _no longer_ calls show or onselect, but only changes the active guid
