### 2020-06-22 c9a6b75
Set default index for map to -1 because the map was not updating newly added markers at index 0. There might be a better way to do this, but this is the first pass solution
