### 2020-06-22 eac5f81
If the activate call is placed before the marker is added to the map, add the marker to the map
