### 2020-06-22 234186e
Consistent naming. newGuid --> newGUID
