### 2020-06-22 92cf167
Set default back to 0 for markers but select will always update to avoid the error after initialization
