+++ src/map/map.ts
@@ -327,7 +327,7 @@ export class PropertiesMap {
         /// Sets the active marker on this map
         /// default of markerData.current is -1 so that currents of value 0 are
         /// still updated
-        if (markerData.current < 0 || indexes.environment !== markerData.current ) {
+        if (markerData.current === 0 || indexes.environment !== markerData.current ) {
             markerData.current = indexes.environment;
             this._updateSelectedMarker(selectedGUID, markerData);
         }
@@ -1451,7 +1451,7 @@ export class PropertiesMap {
      * @param  addedGUID unique string identifier of the marker to add
      * @param  index     numeric index of the structure (with respect to dataset) to show
      */
-    private _addMarker(addedGUID: string, index: number= -1): void {
+    private _addMarker(addedGUID: string, index: number= 0): void {
         if (!this._selected.has(addedGUID)) {
             const activeButton = getByID(`chsp-activate-${addedGUID}`);
             const marker = document.createElement('div');
