### 2020-06-22 3ef122a
`active` is no longer a necessary parameter on Map to enable maps without widgets
