### 2020-06-22 f5f14f4
Removed starterGUID as a necessary parameter for map constructor
