### 2020-06-22 5b75d70
Removed starterGUID as a necessary parameter for map constructor
