+++ src/map/map.ts
@@ -254,12 +254,10 @@ export class PropertiesMap {
      * @param indexer    [[EnvironmentIndexer]] used to translate indexes from
      *                   environments index to structure/atom indexes
      * @param properties properties to be displayed
-     * @param starterGUID if synchronized with a StructureViewer, the GUID of the first structure viewer cell
      */
     constructor(id: string,
                 indexer: EnvironmentIndexer,
                 properties: {[name: string]: Property},
-                starterGUID?: string,
                 ) {
         this._indexer = indexer;
         this.onselect = () => {};
@@ -279,14 +277,6 @@ export class PropertiesMap {
         this._root.appendChild(this._plot);
 
         this._data = new MapData(properties);
-
-        if (starterGUID !== undefined) {
-            this._addMarker(starterGUID);
-            this.active = starterGUID;
-        } else {
-            this._addMarker(generateGUID());
-            this.active = this._selected.keys().next().value;
-        }
 
         this._insertSettingsHTML();
         this._colorReset = getByID<HTMLButtonElement>('chsp-color-reset');
