### 2020-06-22 6d5d521
Improve assertion messages
Instead of showing AssertionError: false === true, show the expression used to generate
the assertion, i.e. AssertionError: tableRoot.tagName.toLowerCase() === 'div'
