### 2020-06-25 9c5fa4b
Rename StructureViewer => ViewerGrid
