+++ src/structure/index.ts
@@ -1,5 +1,5 @@
 /**
- * This module contains the [[StructureViewer|structure viewer]] class and
+ * This module contains the [[ViewersGrid|structure viewer]] class and
  * associated code. Notably, it contains the [[JSmolWidget]] class, that can be
  * used outside of chemiscope as a small wrapper around JSmol.
  *
@@ -8,7 +8,5 @@
  * @preferred
  */
 
-import {StructureViewer} from './structure';
-import {StructurePresets} from './settings';
-
-export {StructureViewer, StructurePresets};
+export {ViewersGrid} from './grid';
+export {StructurePresets} from './settings';
