### 2020-06-25 f8dab8d
Rename StructureViewer => ViewerGrid
