### 2020-06-25 7e8ac25
Move code dealing with structure's HTML settings to settings.ts
