### 2020-06-25 cec2dda
Move code dealing with structure's HTML settings to settings.ts
