+++ src/structure/settings.ts
@@ -3,7 +3,11 @@
  * @module settings
  */
 
-import {HTMLSetting} from '../utils';
+import {HTMLSetting, PositioningCallback} from '../utils';
+import {makeDraggable} from '../utils';
+
+import BARS_SVG from '../static/bars.svg';
+import HTML_SETTINGS from './settings.html';
 
 interface EnvironmentPresets {
     activated: boolean;
@@ -78,7 +82,17 @@ export class StructureSettings {
         bgColor: HTMLSetting<'string'>;
     };
 
-    constructor(presets: Partial<StructurePresets> = {}) {
+    /// The HTML element containing the settings modal
+    private _settingsModal: HTMLElement;
+    // Callback to get the initial positioning of the settings modal.
+    private _positionSettingsModal: PositioningCallback;
+
+    constructor(
+        root: HTMLElement,
+        guid: string,
+        positionSettings: PositioningCallback,
+        presets: Partial<StructurePresets> = {},
+    ) {
         this.bonds = new HTMLSetting('boolean', STRUCTURE_DEFAULTS.bonds);
         this.spaceFilling = new HTMLSetting('boolean', STRUCTURE_DEFAULTS.spaceFilling);
         this.atomLabels = new HTMLSetting('boolean', STRUCTURE_DEFAULTS.atomLabels);
@@ -102,6 +116,12 @@ export class StructureSettings {
             center: new HTMLSetting('boolean', ENVIRONMENTS_DEFAUT.center),
             cutoff: new HTMLSetting('number', ENVIRONMENTS_DEFAUT.cutoff),
         };
+
+        this._positionSettingsModal = positionSettings;
+
+        this._settingsModal = this._insertHTML(root, guid);
+        document.body.appendChild(this._settingsModal);
+        this._bindSettings(guid);
 
         this.applyPresets(presets);
     }
@@ -160,5 +180,102 @@ export class StructureSettings {
             supercell: [ this.supercell[0].value, this.supercell[1].value, this.supercell[2].value],
             unitCell: this.unitCell.value,
         };
+    }
+
+    /**
+     * Remove all HTML added by this [[StructureSettings]] in the current
+     * document
+     */
+    public remove(): void {
+        this._settingsModal.remove();
+    }
+
+    /**
+     * Insert HTML needed for setting in the page, adding the "open settings"
+     * button to `root`. The setting modal HTML element is returned, not yet
+     * inserted in the document.
+     *
+     * @param  root where to place the HTML button
+     * @param  guid unique identifier of the corresponding JSmolWidget, used as
+     *              prefix for all elements ID
+     * @return      the HTML element containing the setting modal
+     */
+    private _insertHTML(root: HTMLElement, guid: string): HTMLElement {
+        // use HTML5 template to generate a DOM object from an HTML string
+        const template = document.createElement('template');
+        template.innerHTML = `<button
+            class="btn btn-light btn-sm chsp-viewer-button"
+            data-target="#${guid}-settings"
+            data-toggle="modal"
+            style="top: 5px; right: 5px; opacity: 1;">
+                <div>${BARS_SVG}</div>
+            </button>`;
+        const openSettings = template.content.firstChild!;
+        root.append(openSettings);
+
+        // replace id to ensure they are unique even if we have mulitple viewers
+        // on a single page
+        template.innerHTML = HTML_SETTINGS
+            .replace(/id=(.*?) /g, (_: string, id: string) => `id=${guid}-${id} `)
+            .replace(/for=(.*?) /g, (_: string, id: string) => `for=${guid}-${id} `)
+            .replace(/data-target=#(.*?) /g, (_: string, id: string) => `data-target=#${guid}-${id} `);
+
[…]
