+++ src/structure/widget.ts
@@ -7,12 +7,9 @@ import assert from 'assert';
 
 import {JmolObject, JSmolApplet} from 'jsmol';
 
-import {generateGUID, getByID, makeDraggable} from '../utils';
+import {generateGUID, getByID} from '../utils';
 import {PositioningCallback} from '../utils';
 import {StructurePresets, StructureSettings} from './settings';
-
-import BARS_SVG from '../static/bars.svg';
-import HTML_SETTINGS from './settings.html';
 
 // tslint:disable-next-line: no-var-requires
 require('../static/chemiscope.css');
@@ -45,22 +42,22 @@ function createStyleSheet(rules: string[]): CSSStyleSheet {
  * cells both on negative and positive sides.
  */
 function supercell_555(supercell: [number, number, number]): string {
-    const supercellCenter = [
+    const central = [
         Math.floor(supercell[0] / 2),
         Math.floor(supercell[1] / 2),
         Math.floor(supercell[2] / 2),
     ];
 
     const first = [
-        5 - supercellCenter[0],
-        5 - supercellCenter[1],
-        5 - supercellCenter[2],
+        5 - central[0],
+        5 - central[1],
+        5 - central[2],
     ];
 
     const second = [
-        5 + supercell[0] - (supercellCenter[0] + 1),
-        5 + supercell[1] - (supercellCenter[1] + 1),
-        5 + supercell[2] - (supercellCenter[2] + 1),
+        5 + supercell[0] - (central[0] + 1),
+        5 + supercell[1] - (central[1] + 1),
+        5 + supercell[2] - (central[2] + 1),
     ];
 
     return `{${first[0]}${first[1]}${first[2]} ${second[0]}${second[1]}${second[2]} 1}`;
@@ -130,14 +127,12 @@ export class JSmolWidget {
 
     /// The HTML element serving as root element for the viewer
     private _root: HTMLElement;
-    /// The HTML element containing the settings modal
-    private _settingsModal!: HTMLElement;
     /// Reference to the global Jmol variable
     private _Jmol: JmolObject;
     /// Reference to the JSmol applet to be updated with script calls
-    private _applet!: JSmolApplet;
+    private _applet: JSmolApplet;
     /// Representation options from the HTML side
-    private _settings!: StructureSettings;
+    private _settings: StructureSettings;
     /// The supercell used to intialize the viewer
     private _initialSupercell?: [number, number, number];
     // button to reset the environment cutoff to its original value
@@ -216,9 +211,14 @@ export class JSmolWidget {
         this._cellInfo.classList.add('chsp-cell-info', 'chsp-hide-if-no-cell', 'badge', 'badge-light');
         this._root.appendChild(this._cellInfo);
 
-        this._createOptions();
-        this._createApplet(j2sPath, serverURL);
+        this._settings = new StructureSettings(
+            this._root,
+            this.guid,
+            (rect) => this.positionSettingsModal(rect),
+        );
+        this._connectSettings();
 
+        this._applet = this._createApplet(j2sPath, serverURL);
         if (!isSafari()) {
             // Invert (cf -1 below) wheel zoom direction to match the one in the
             // map. _DELTAY is replaced by 1 or -1 depending on the wheel/scroll
@@ -290,7 +290,7 @@ export class JSmolWidget {
         if (this._root.parentElement !== null) {
             this._root.parentElement.innerHTML = '';
         }
-        this._settingsModal.remove();
+        this._settings.remove();
     }
 
     /**
@@ -388,7 +388,7 @@ export class JSmolWidget {
             keepOrientation = options.keepOrientation;
         }
 
-        const trajectoryOptions = getByID<HTMLElement>(`${this.guid}-trajectory-settings-group`);
[…]
