### 2020-06-25 #2117 @Luthaf
@cMadan, what is the next step here? I've seen we have to provide a zenodo DOI for a fixed release, is there anything else to do on our side?
