### 2020-06-26 be32746
Update src/map/map.ts
Co-authored-by: Guillaume Fraux <luthaf@luthaf.fr>
