### 2020-06-26 3c58f93
Added comments describing new connecting functions (onremove and activate)
