### 2020-06-26 4fdb230
Added comments describing new connecting functions (onremove and activate)
