### 2020-06-26 e70f4bb
Added description of  to map.select
