### 2020-06-26 54c3629
Broke up if/else block in map.get active
