### 2020-06-26 2867ea6
Broke up if/else block in map.get active
