### 2020-06-26 28bdc2d
Separated *hiding* markers (i.e. from 2D --> 3D) from *removing* markers (when deleting a point) from the map
