### 2020-06-26 c5f11f6
Removing artifact assertion
