### 2020-06-26 7a4cbc1
Double check grid has been initialized
