### 2020-06-26 748a30b
Do not allow removeWidget to remove the last widget
