### 2020-06-26 #2117 @Luthaf
Do you mean https://github.com/cosmo-epfl/chemiscope/issues/21? I think so, the goal is to introduce automated testing to the code, in addition to current manual checking. It is not totally trivial to do, since the project is mostly a graphical interface, which is hard to unit test.

We discussed this with @mkhorton starting https://github.com/openjournals/joss-reviews/issues/2117#issuecomment-625103628, and I think both reviewer agreed that automated tests are not required as part of the review.
