### 2020-06-26 cc09e24
Some small edits
