### 2020-06-26 #2117 @mkhorton
Yes, I definitely would have liked to see automated tests, but our reading of the JOSS guidelines was that this was _not_ required to fulfill the requirements of the review (though of course encouraged!).
