### 2020-06-26 #2117 @kthyng
That is correct that automated testing is not required, and I read above that *a* means of verifying correct behavior has been included. Thank you @mkhorton!
