### 2020-06-26 #2117 @kthyng
@Luthaf I made some small changes to your paper in [this PR](https://github.com/cosmo-epfl/chemiscope/pull/54), please consider and merge if you agree with them. I looked through your references too and they look reasonably capitalized, etc, but please verify too.
