### 2020-06-26 #2117 @kthyng
@Luthaf after the paper is finalized, please verify the current version of the code. Also, go to a place like Zenodo and create an archive of your code and report back here with the doi. Be sure that the metadata for the Zenodo archive exactly matches that of your JOSS paper (title and authors) — you may need to update it there. This is a request but not mandatory.
