### 2020-06-26 #2117 @Luthaf
We just set up zenodo, and released version 0.2.0 this morning, corresponding to the latest master (minus your PR): https://zenodo.org/record/3909400

The title should match already match, the authors also include two more people, one who helped correct a typo, and @rosecers. @rosecers made some significant contributions to the code since the paper was submited, so we'll discuss co-authorship on the paper. I'll have to correct the zenodo author list anyway, I'll try to do this tomorrow.
