### 2020-06-30 #2117 @Luthaf
@cMadan, after discussion I've added @rosecers as co-author of the paper, and added the corresponding funding body to the acknowledgements, so this should be done from our side!
