### 2020-07-01 136dbed
Generic settings presets application & saving
Instead of multiplying interfaces, use introspection to dump & apply presets
from a given setting group
