+++ src/map/settings.ts
@@ -3,10 +3,9 @@
  * @module settings
  */
 
-import assert from 'assert';
-
-import {HTMLSetting, PositioningCallback} from '../utils';
-import {getByID, makeDraggable, settingsValidator} from '../utils';
+import {HTMLSetting, SettingsGroup, SettingsPreset} from '../settings';
+import {settingsValidator} from '../settings';
+import {getByID, makeDraggable, PositioningCallback} from '../utils';
 import {NumericProperties} from './data';
 
 import {COLOR_MAPS} from './colorscales';
@@ -14,66 +13,23 @@ import {COLOR_MAPS} from './colorscales';
 import BARS_SVG from '../static/bars.svg';
 import HTML_SETTINGS from './settings.html';
 
-export interface AxisPresets {
-    /// Which property should we use for this axis
-    property: string;
-    /// Which scale (linear/log) should we use
-    scale: string;
-    /// The minimal value for this axis
-    min: number;
-    /// The maximal value for this axis
-    max: number;
-}
-
-const AXIS_DEFAULTS: AxisPresets = {
-     max: 0,
-     min : 0,
-     property : '',
-     scale : 'linear',
-};
-
 /** HTML element holding settings for a given axis (x, y, z, color) */
-export class AxisSettings {
+export class AxisSettings extends SettingsGroup {
     public property: HTMLSetting<'string'>;
     public scale: HTMLSetting<'string'>;
     public min: HTMLSetting<'number'>;
     public max: HTMLSetting<'number'>;
 
     constructor(validProperties: string[]) {
-        this.max = new HTMLSetting('number', AXIS_DEFAULTS.max);
-        this.min = new HTMLSetting('number', AXIS_DEFAULTS.min);
-        this.property = new HTMLSetting('string', AXIS_DEFAULTS.property);
-        this.scale = new HTMLSetting('string', AXIS_DEFAULTS.scale);
+        assert(validProperties.length > 0);
+        super();
+        this.max = new HTMLSetting('number', 0);
+        this.min = new HTMLSetting('number', 0);
+        this.property = new HTMLSetting('string', validProperties[0]);
+        this.scale = new HTMLSetting('string', 'linear');
 
         this.property.validate = settingsValidator(validProperties, 'axis');
         this.scale.validate = settingsValidator(['linear', 'log'], 'axis scale');
-    }
-
-    /**
-     * Applies presets, possibly filling in with default values
-     */
-    public applyPresets(presets: Partial<AxisPresets> = {}): void {
-        const initial: AxisPresets = {
-            ...AXIS_DEFAULTS,
-            ...presets,
-        };
-
-        this.max.value = initial.max;
-        this.min.value = initial.min;
-        this.property.value = initial.property;
-        this.scale.value = initial.scale;
-    }
-
-    /**
-     * Dumps presets, in a way that can e.g. be serialized to json
-     */
-    public dumpPresets(): AxisPresets {
-        return {
-            max: this.max.value,
-            min: this.min.value,
-            property: this.property.value,
-            scale: this.scale.value,
-        };
     }
 
     /** Disable auxiliary settings (min/max/scale) related to this axis */
@@ -91,34 +47,7 @@ export class AxisSettings {
     }
 }
 
-export interface MapPresets {
-    x: Partial<AxisPresets>;
-    y: Partial<AxisPresets>;
-    z: Partial<AxisPresets>;
-    color: Partial<AxisPresets>;
-    palette: string;
-    symbol: string;
-    size: {
-        property?: string,
-        factor?: number,
-    };
-}
-
[…]
