+++ src/structure/settings.ts
@@ -3,53 +3,13 @@
  * @module settings
  */
 
-import {HTMLSetting, PositioningCallback} from '../utils';
-import {makeDraggable} from '../utils';
+import {HTMLSetting, SettingsGroup, SettingsPreset} from '../settings';
+import {makeDraggable, PositioningCallback} from '../utils';
 
 import BARS_SVG from '../static/bars.svg';
 import HTML_SETTINGS from './settings.html';
 
-interface EnvironmentPresets {
-    activated: boolean;
-    center: boolean;
-    cutoff: number;
-    bgStyle: string;
-    bgColor: string;
-}
-
-export interface StructurePresets {
-    bonds: boolean;
-    spaceFilling: boolean;
-    atomLabels: boolean;
-    unitCell: boolean;
-    packedCell: boolean;
-    supercell: number[];
-    rotation: boolean;
-    axes: string;
-    environments: Partial<EnvironmentPresets>;
-    keepOrientation: boolean;
-}
-
-export const STRUCTURE_DEFAULTS: StructurePresets = {
-    atomLabels: false,
-    axes: 'off',
-    bonds: true,
-    environments: {
-        activated: true,
-        bgColor: 'grey',
-        bgStyle: 'licorice',
-        center: false,
-        cutoff: 4.0,
-    },
-    keepOrientation: false,
-    packedCell: false,
-    rotation: false,
-    spaceFilling: false,
-    supercell: [1, 1, 1],
-    unitCell: false,
-};
-
-export class StructureSettings {
+export class StructureSettings extends SettingsGroup {
     // should we show bonds
     public bonds: HTMLSetting<'boolean'>;
     // should we use space filling representation
@@ -91,30 +51,32 @@ export class StructureSettings {
         root: HTMLElement,
         guid: string,
         positionSettings: PositioningCallback,
-        presets: Partial<StructurePresets> = {},
+        presets: SettingsPreset = {},
     ) {
-        this.bonds = new HTMLSetting('boolean', STRUCTURE_DEFAULTS.bonds);
-        this.spaceFilling = new HTMLSetting('boolean', STRUCTURE_DEFAULTS.spaceFilling);
-        this.atomLabels = new HTMLSetting('boolean', STRUCTURE_DEFAULTS.atomLabels);
-        this.unitCell = new HTMLSetting('boolean', STRUCTURE_DEFAULTS.unitCell);
-        this.packedCell = new HTMLSetting('boolean', STRUCTURE_DEFAULTS.packedCell);
+        super();
+
+        this.bonds = new HTMLSetting('boolean', true);
+        this.spaceFilling = new HTMLSetting('boolean', false);
+        this.atomLabels = new HTMLSetting('boolean', false);
+        this.unitCell = new HTMLSetting('boolean', false);
+        this.packedCell = new HTMLSetting('boolean', false);
+        this.rotation = new HTMLSetting('boolean', false);
+
         this.supercell = [
-            new HTMLSetting('int', STRUCTURE_DEFAULTS.supercell[0]),
-            new HTMLSetting('int', STRUCTURE_DEFAULTS.supercell[1]),
-            new HTMLSetting('int', STRUCTURE_DEFAULTS.supercell[2]),
+            new HTMLSetting('int', 1),
+            new HTMLSetting('int', 1),
+            new HTMLSetting('int', 1),
         ];
 
-        this.rotation = new HTMLSetting('boolean', STRUCTURE_DEFAULTS.rotation);
-        this.axes = new HTMLSetting('string', STRUCTURE_DEFAULTS.axes);
-        this.keepOrientation = new HTMLSetting('boolean', STRUCTURE_DEFAULTS.keepOrientation);
[…]

[…]
