### 2020-07-01 3f4d313
Add settings validation for the structures
