### 2020-07-01 6d24664
Replace internal errors with assert
