### 2020-07-01 e9ecdf1
Replace internal errors with assert
