### 2020-07-01 8c11a56
Incorporate @ceriottm final feedback
