+++ paper/paper.md
@@ -25,25 +25,26 @@ bibliography: paper.bib
 
 # Summary
 
-The number of materials or molecules that can be created by combining
-different chemical elements in various proportions and spatial arrangements is
-enormous. Computational chemistry can be used to generate databases containing
-billions of potential structures [@Ruddigkeit2012], and predict some of the
-associated properties [@Montavon2013; @Ramakrishnan2014]. Unfortunately, the
-very large number of structures makes exploring such database — to understand
+The number of materials or molecules that can be created by combining different
+chemical elements in various proportions and spatial arrangements is enormous.
+Computational chemistry can be used to generate databases containing billions of
+potential structures [@Ruddigkeit2012], and predict some of the associated
+properties [@Montavon2013; @Ramakrishnan2014]. Unfortunately, the very large
+number of structures makes exploring such database — to understand
 structure-property relations or find the *best* structure for a given
-application — a daunting task. In the recent years, multiple molecular
-*descriptors* [@Behler2007; @Bartok2013; @Willatt2019] have been developed to
-compute structural similarities between materials or molecules, incorporating
-physically-relevant information and symmetries. These descriptors can be used
-for unsupervised machine learning applications, such as clustering or
-classification of the different structures, and high-throughput screening of
-database for specific properties [@Maier2007; @De2017; @Hautier2019].
-Unfortunately, the dimensionality of most descriptors is very high, which makes
-the resulting classifications, clustering or mapping very hard to visualize.
-Additional dimensionality reduction algorithm [@Schlkopf1998; @Ceriotti2011;
-@McInnes2018] can reduce the number of relevant dimensions to a handful,
-creating 2D or 3D maps of the full database.
+application — a daunting task. In recent years, multiple molecular
+*representations* [@Behler2007; @Bartok2013; @Willatt2019] have been developed
+to compute structural similarities between materials or molecules, incorporating
+physically-relevant information and symmetries. The features associated with
+these representations can be used for unsupervised machine learning
+applications, such as clustering or classification of the different structures,
+and high-throughput screening of database for specific properties [@Maier2007;
+@De2017; @Hautier2019]. Unfortunately, The dimensionality of these features (as
+well as most of other descriptors used in chemical and materials informatics) is
+very high, which makes the resulting classifications, clustering or mapping very
+hard to visualize. Additional dimensionality reduction algorithm
+[@Schlkopf1998; @Ceriotti2011; @McInnes2018] can reduce the number of relevant
+dimensions to a handful, creating 2D or 3D maps of the full database.
 
 ![The Qm7b database [@Montavon2013] visualized with chemiscope](screenshot.png)
 
@@ -55,11 +56,11 @@ point corresponds to a chemical entity. The axes, color, and style of each point
 can be set to represent a property or a structural descriptor to visualize
 structure-property relations directly. Structural descriptors are not computed
 directly by chemiscope, but must be obtained from one of the many codes
-implementing such descriptors [@librascal; @QUIP]. Since the most common
+implementing general-purpose atomic representation [@librascal; @QUIP] or more specialized descriptors. Since the most common
 descriptors can be very high dimensional, it can be convenient to apply a
 dimensionality reduction algorithm that maps them to a lower-dimensional space
 for easier visualization. For example the sketch-map algorithm [@Ceriotti2011]
-was used with the Smooth Overlap of Atomic Positions descriptor [@Bartok2013] to
+was used with the Smooth Overlap of Atomic Positions representation [@Bartok2013] to
 generate the visualization in Figure 1. The right panel displays the
 three-dimensional structure of the chemical entities, possibly including
 periodic repetition for crystals. Visualizing the chemical structure can help to
@@ -99,14 +100,14 @@ slower, while still handling 100k points easily.
 The use of web technologies makes chemiscope usable from different operating
[…]
