### 2020-07-01 #2117 @Luthaf
We've made some tiny phrasing changes & improved the use of descriptor vs representation: https://github.com/cosmo-epfl/chemiscope/commit/8c11a56929c4793a845931939c7e60bfde740493
