### 2020-07-01 #2117 @danielskatz
Please see https://github.com/cosmo-epfl/chemiscope/pull/55

Also, is "Additional" in the first paragraph needed?
And the caption of figure 2 seems to end abruptly.
