### 2020-07-01 2f76a6d
More minor changes to the paper (#55)
