+++ paper/paper.md
@@ -39,19 +39,19 @@ physically-relevant information and symmetries. The features associated with
 these representations can be used for unsupervised machine learning
 applications, such as clustering or classification of the different structures,
 and high-throughput screening of database for specific properties [@Maier2007;
-@De2017; @Hautier2019]. Unfortunately, The dimensionality of these features (as
+@De2017; @Hautier2019]. Unfortunately, the dimensionality of these features (as
 well as most of other descriptors used in chemical and materials informatics) is
 very high, which makes the resulting classifications, clustering or mapping very
-hard to visualize. Additional dimensionality reduction algorithm
-[@Schlkopf1998; @Ceriotti2011; @McInnes2018] can reduce the number of relevant
-dimensions to a handful, creating 2D or 3D maps of the full database.
+hard to visualize. Dimensionality reduction algorithms [@Schlkopf1998;
+@Ceriotti2011; @McInnes2018] can reduce the number of relevant dimensions to a
+handful, creating 2D or 3D maps of the full database.
 
 ![The Qm7b database [@Montavon2013] visualized with chemiscope](screenshot.png)
 
 Chemiscope is a graphical tool for the interactive exploration of materials and
 molecular databases, correlating local and global structural descriptors with
-the physical properties of the different systems. The interface is composed by
-two panels. The left panel consists of a 2D or 3D scatter plot, in which each
+the physical properties of the different systems. The interface consists of
+two panels. The left panel displays a 2D or 3D scatter plot, in which each
 point corresponds to a chemical entity. The axes, color, and style of each point
 can be set to represent a property or a structural descriptor to visualize
 structure-property relations directly. Structural descriptors are not computed
@@ -63,8 +63,8 @@ for easier visualization. For example the sketch-map algorithm [@Ceriotti2011]
 was used with the Smooth Overlap of Atomic Positions representation [@Bartok2013] to
 generate the visualization in Figure 1. The right panel displays the
 three-dimensional structure of the chemical entities, possibly including
-periodic repetition for crystals. Visualizing the chemical structure can help to
-find an intuitive rationalization of the layout of the dataset and the
+periodic repetition for crystals. Visualizing the chemical structure can help
+in finding an intuitive rationalization of the layout of the dataset and the
 structure-property relations.
 
 Whereas similar tools [@Gong2013; @Gutlein2014; @Probst2017; @ISV] only allow
@@ -78,7 +78,7 @@ machine-learning schemes that decompose properties into atom-centred
 contributions [@Behler2007; @Bartok2010].
 
 
-![Database of chemical shieldings [@Paruzzo2018] in chemiscope demonstrating the use of a 3D plot and atomic environments highlighting](./screenshot-3d.png)
+![Database of chemical shieldings [@Paruzzo2018] in chemiscope demonstrating the use of a 3D plot and highlighting of atomic environments](./screenshot-3d.png)
 
 Chemiscope took strong inspiration from a previous similar graphical software,
 the interactive sketch-map visualizer [@ISV]. This previous software was used in
@@ -89,7 +89,7 @@ databases, and the mapping of structure-property relationships [@De2016;
 # Implementation
 
 Chemiscope is implemented using the web platform: HTML5, CSS and WebGL to
-display graphical elements, and TypeScript (compiled to JavaScript) for the
+display graphical elements, and TypeScript (compiled to JavaScript) for
 interactivity. It uses [Plotly.js](https://plot.ly/javascript/) to render and
 animate 2D and 3D plots; and the JavaScript version of [Jmol](http://jmol.org/)
 to display atomic structures. The visualization is fast enough to be used with
@@ -100,7 +100,7 @@ slower, while still handling 100k points easily.
 The use of web technologies makes chemiscope usable from different operating
[…]
