### 2020-07-01 #2117 @Luthaf
Thanks for the feedback, I hope I addressed your comments in https://github.com/cosmo-epfl/chemiscope/pull/55/commits/a9828421d8832537a3ff4992f2a7de5cb3ef8fd2
